### 2026-01-15 #743 @ErickNavarroD
RAMEN: Regional Association of Methylome variability with the Exposome and geNome
Submitting Author Name: Erick I. Navarro-Delgado
Submitting Author Github Handle: @ErickNavarroD
Repository:  https://github.com/ErickNavarroD/RAMEN
Version submitted: 2.0.0
Submission type: Standard

Version accepted: TBD

---

```
Package: RAMEN
Title: RAMEN: Regional Association of Methylome variability with the Exposome and geNome
Version: 2.0.0
    person("Erick I.", "Navarro-Delgado", , "ericknadel98@hotmail.com", role = c("aut", "cre"),
Description: R package that identifies which genetic (G), environmental (E), additive (G+E) or interaction (GxE) effect better explains DNA methylation levels in Variable Methylated Loci using microarray data.
    BiocStyle,
    doRNG,
    foreach,
    GenomicRanges,
    glmnet,
    IlluminaHumanMethylation450kanno.ilmn12.hg19,
    IlluminaHumanMethylationEPICanno.ilm10b4.hg19,
    IlluminaHumanMethylationEPICv2anno.20a1.hg38,
    IRanges,
    iterators,
    lifecycle,
    relaimpo,
    S4Vectors,
    tibble
```

The objective of the package is to provide a FAIR framework to analyze DNA methylation microarray, genotype and exposome data, and conduct a series of operations to model the association between DNA methylation and both genetic and environmental data. The RAMEN package puts together functions from different already-developed statistical packages to conduct the modelling, and is designed to conduct a specific analysis within the field of epigenetics.

The target audience of this package is researchers in the epigenetics field that have access to epidemiological cohorts with genetic, environmental and DNA methylation data. The main application of this package is to model the individual, additive and interactive contributions of genetics and environmental differences to genome-wide DNA methylation levels in a population.

To my knowledge there are no other R packages that accomplish the same thing.

Yes.

This package is within the scope of rOpenSci, as the process of submitting RAMEN to peer review is my project for the rOpenSci Champions 2025 program.
