### 2026-01-20 #743 @jhollist
@ErickNavarroD thank you for the submission.

Things are looking good, but i do have a couple of thoughts about category for the package and stats vs regular submission.

This could possibly fit as a stats package (https://stats-devguide.ropensci.org/), likely as a Regression and Supervised Learning package.  You would need to comply with at least half of the Regression and Supervised Learning Standards (https://stats-devguide.ropensci.org/standards.html#standards-regression), plus general standards (https://stats-devguide.ropensci.org/standards.html#standards-general).  If you are able to do that, then it's a stats package.

If you think this is better as a standard package, I would look at the Field and Lab Reproducibility Tools (https://devguide.ropensci.org/softwarereview_policies.html#package-categories).

Could you take a read through the links and let me know your thoughts.  The process and checks for stats packages are a little different, but we can assist with that process if you want to go that route.
