### 2026-02-05 #743 @ErickNavarroD
Thank you for your reply @jhollist !
I appreciate your thoughts to find the right category for the package. I went through the scope of both the Stats and Field and Reproducibility Lab tools streams, and I think the package would fit better in the later one.
The RAMEN package is closer to a workflow that puts together functions from different already-developed statistical packages to conduct the modelling, and is designed to conduct a specific analysis within the field of epigenetics. The objective of the package is to provide a FAIR framework that outputs summary statistics and metrics of models in multiple sites of the genome (~200k), rather than developing new statistical algorithms to obtain a model object that can be used for prediction. So, even though my package exists within the Regression framework and uses their methods, I think it is best suited as a Lab Reproducibility Tool.  Please let me know your thoughts about this.
Thanks!
