### 2026-02-06 #743 @jhollist
@ErickNavarroD Thanks for the response!  Just wanted to let you know that @emilyriederer has taken over as EiC.  She's got RAMEN on her radar and will likely respond soon.

Thanks again for your submission.
