### 2026-02-09 #743 @maurolepore
@ErickNavarroD, to help you track my comments I'll tag them with "ml" and a numbered sequence, e.g. ml01, ml02, and so on:

* Comments following bullets are for you to consider -- you may or may not respond to them.
- [ ] Comments following check-boxes are requests for some action -- please respond.

For example, here are my first two comments, and both need a response:

- [ ] ml01. If you agree, can you please confirm the category on the initial comment? (I see that the one checked is "data munging").
- [ ] ml02. Could you please suggest three reviewers without [conflicts of interest](https://devguide.ropensci.org/softwarereview_policies.html#coi)? We may use at most one, but the having a few more helps us better understand the type of expertise you think would be beneficial. This may be hard work but helps find good reviewers as quickly as possible.
