### 2026-02-17 #743 @maurolepore
@ErickNavarroD, thanks again for sharing your work with rOpenSci.

Overall the package seems in great shape and many of my comments are optional. The comments marked with check-boxes require some action (and it's okay to disagre and explain why). The intent is to speed up the review process by flagging now some issues that, if addressed now, will help reviewers focus on the most interesting aspects of the package.

*I used AI tools ([Claude Code](https://claude.com/product/claude-code), model: claude-opus-4-6) to research the evidence for these checks. I confirmed the evidence, decided what to include, and marked each as optional or required. The interpretation, judgment and responsibility are mine.*

      - [ ] ml03. In README I see what RAMEN does. Please also add why it's useful, particularly in comparison with other related packages (see also fit and overlap below). The discussion about the category in which this package fits may help frame its usefulness.
      - [ ] ml04. In README I see the Usage section points to a rich vignette. Also please add a basic example to help understand some representative inputs and outputs as quickly as possible.
      - ml05. In README I see the section "How to get help for RAMEN" mentions GitHub issues but provides no link. Consider adding the link or relying on other standard places (e.g. the website) that automatically include that link.
      - [ ] ml06. In README I see authors not listed in DESCRIPTION. Please review for consistency.
      - ml07. If some grouping makes sense, consider using it over the default alphabetical order. I think this is optional but may become more important if the number of functions grows.
  - [ ] ml08. Please acknowledge the related packages and explain how this package is still useful/necessary.
  - [ ] ml09. Please try the installation instructions in a fresh environment to identify requirements beyond what's documented in README, e.g. Bioconductor packages.
  - [ ] ml10. Please review `.github/CONTRIBUTING.md`. I see the placeholder "our_package" and the phrase "this and other tidyverse packages".

---

- [ ] ml11. On the website I see some functions and datasets do not print the output. Please show a bit of output, e.g. with `str()` or `head()`, to help readers understand what to expect even before they run the code themselves. See ([building.html#general](https://devguide.ropensci.org/building.html#general)): *All exported package functions should be fully documented with examples.*
- [ ] ml12. In DESCRIPTION I see an email address that is not the single one used across the package documentation. Please review for consistency.
- [ ] ml13. Please add `.Rbuildignore` to avoid building the tarball with irrelevant files and directories, e.g. data-raw/. You may do this automatically with `usethis::use_data_raw()`
- ml14. Consider removing unused .png files if any, and removing docs/ if the website is built on GitHub Actions.
- ml15. Please double check the citation follows the ([dev guide requirements](https://devguide.ropensci.org/building.html#citation-file))
- ml16. DESCRIPTION declares `Depends: R (>= 4.2.0)` but CI (R-CMD-check.yaml) only tests `release`, `devel`, `oldrel-1` (~R 4.4). Please consider extending the CI workflow to cover the stated R-version dependency.
- ml17. Consider structuring R/ and tests/ files so that they play well with `usethis::use_r()` and `usethis::use_test()`. Beyond the convenience, using a standard structure helps reviewers and contributors navigate an unfamiliar package.
- ml18. Consider rewriting tests so that they are [self-contained tests](https://r-pkgs.org/testing-design.html#sec-testing-design-self-contained). Ideally separate tests should not share state. Tests are easier to understand, debug, and maintain when they avoid code outside `test_that()`, i.e. no calls to `library()` (but see `withr::with_namespace()` and `withr::with_package()`) nor data shared across tests. See also [testthat special files](https://testthat.r-lib.org/articles/special-files.html), the [fixtures vignette](https://testthat.r-lib.org/articles/test-fixtures.html), the [R Packages testing design](https://r-pkgs.org/testing-design.html#sec-testing-design-principles), and [DRY, DAMP, or WET?](https://ropensci.org/blog/2024/03/18/dry-damp/).
- [ ] ml19. Please review the test runtime and explore ways to make tests run in a few seconds rather than a few minutes. The faster the tests the more often we run them and the more bugs we catch :-)
- [ ] ml20. Please review tests so that the test-report prints cleanly. Busy output can hide important details and confuse reviewers.

<details><summary>Test output (devtools::test())</summary>

```
ℹ Testing RAMEN
  Setting options('download.file.method.GEOquery'='auto')
  Setting options('GEOquery.inmemory.gpl'=FALSE)
           __      _             ___
           )_)    /_)    )\/)    )_     )\  )
          / \    / /    (  (    (__    (  \(

                    (   )  (  (
                   (  ( )  (  )
                     )    )  (
                  _.(--'(''--.._
                 /, _..-----).._,\
                |  `'''-----'''`  |
                 \               /
                  '.           .'
                    '--.....--'

  If you use RAMEN for your analysis, please cite: Navarro-Delgado, E.I.,
  Czamara, D., Edwards, K. et al. RAMEN: Dissecting individual, additive
  and interactive gene-environment contributions to DNA methylome variability
  in cord blood. Genome Biol 26, 421 (2025).
  https://doi.org/10.1186/s13059-025-03864-4
  WARNING: 'test-workflow.R:6:1' --------------------
  package 'dplyr' was built under R version 4.5.2
      ▆
   1. └─base::library(dplyr) at test-workflow.R:6:1
   2.   └─base (local) testRversion(pkgInfo, package, pkgpath)

  Attaching package: 'dplyr'

  The following objects are masked from 'package:stats':

      filter, lag

  The following objects are masked from 'package:base':

      intersect, setdiff, setequal, union

  Identifying Highly Variable Probes...
  Identifying sparse Variable Methylated Probes
  Identifying Variable Methylated Regions...
  Applying correlation filter to Variable Methylated Regions...
  WARNING: 'test-workflow.R:9:1' --------------------
  executing %dopar% sequentially: no parallel backend registered
      ▆
   1. ├─RAMEN::findVML(...) at test-workflow.R:9:1
   2. │ └─... %>% ... at RAMEN/R/findVML.R:262:5
   3. ├─GenomicRanges::makeGRangesFromDataFrame(., keep.extra.columns = TRUE)
   4. ├─dplyr::filter(., median_correlation > cor_threshold)
   5. └─RAMEN::medCorVMR(VMR_df = ., methylation_data = methylation_data)
   6.   └─... %dopar% ... at RAMEN/R/medCorVMR.R:41:3
   7.     └─foreach:::getDoPar()

  Identifying Highly Variable Probes...
  Identifying sparse Variable Methylated Probes
  Identifying Variable Methylated Regions...
  Applying correlation filter to Variable Methylated Regions...
  Identifying Highly Variable Probes...
  Identifying sparse Variable Methylated Probes
  Identifying Variable Methylated Regions...
  Applying correlation filter to Variable Methylated Regions...
  Identifying Highly Variable Probes...
  Identifying sparse Variable Methylated Probes
  Identifying Variable Methylated Regions...
  Applying correlation filter to Variable Methylated Regions...
  Reminder: please make sure that the positions of the VML data frame and the ones in the genotype information are from the same genome build.
  Reminder: please make sure that the positions of the VML data frame and the ones in the genotype information are from the same genome build.
  Reminder: please make sure that the positions of the VML data frame and the ones in the genotype information are from the same genome build.
  Reminder: please make sure that the positions of the VML data frame and the ones in the genotype information are from the same genome build.
  Reminder: please make sure that the positions of the VML data frame and the ones in the genotype information are from the same genome build.
  Reminder: please make sure that the positions of the VML data frame and the ones in the genotype information are from the same genome build.
  WARNING: 'test-workflow.R:363:1' ------------------
  package 'doRNG' was built under R version 4.5.2
      ▆
   1. └─RAMEN::selectVariables(...) at test-workflow.R:363:1
   2.   ├─... %dorng% ... at RAMEN/R/selectVariables.R:148:3
   3.   │ └─base::do.call(`%dopar%`, list(obj, ex), envir = parent.frame())
   4.   └─foreach (local) `<fn>`(...)
   5.     └─e$fun(obj, substitute(ex), parent.frame(), e$data)
   6.       └─base::library(p, character.only = TRUE)
   7.         └─base (local) testRversion(pkgInfo, package, pkgpath)

  Loading required package: foreach
  Loading required package: rngtools
  Starting permutation 1 of 2
  Starting variable selection of permutation 1 of 2
  Starting lmGE in permutation 1 of 2
  Wrapping up permutation 1 of 2
  Starting permutation 2 of 2
  Starting variable selection of permutation 2 of 2
  Starting lmGE in permutation 2 of 2
  Wrapping up permutation 2 of 2
  Starting permutation 1 of 2
  Starting variable selection of permutation 1 of 2
  Starting lmGE in permutation 1 of 2
  Wrapping up permutation 1 of 2
  Starting permutation 2 of 2
  Starting variable selection of permutation 2 of 2
  Starting lmGE in permutation 2 of 2
  Wrapping up permutation 2 of 2
  Starting permutation 1 of 2
  Starting variable selection of permutation 1 of 2
  Starting permutation 2 of 2
  Starting variable selection of permutation 2 of 2
  package 'testthat' was built under R version 4.5.2
```

- ml21. I see both `%>%` and `|>`. Consider using one type of pipe consistently.
- ml22. Consider finding opportunities to deduplicate code (e.g. `c(list(NULL), list("")`) and simplify logic.
- ml23. In `map_revmap_names` consider removing `\dontrun{}` since it seems unnecessary.
- ml24. In the output of pkgcheck I see "Only 1 non-exported function vs 8 exported. High cyclocomplexity + low non-exported function count suggests core functions are large monoliths that could be broken into smaller internal helpers."
- [ ] ml25. Please style the code with an automated tool like `air format .` for readability. Using a standard code style helps reviewers and contributors more easily understand the code.
- [ ] ml26. Please address the results of `urlchecker::url_check()`

<details><summary>urlchecker</summary>

```r
urlchecker::url_check()
! Warning: vignettes/RAMEN.Rmd:261:211
Moved
Please make sure that your data has no NAs, since the LASSO implementation we use in RAMEN does not support missing values, and that all values are numeric. If your data has missing values, consider [handling](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3668100/) them.
                                                                                                                                                                                                                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
                                                                                                                                                                                                                  https://pmc.ncbi.nlm.nih.gov/articles/PMC3668100/
! Warning: vignettes/RAMEN.Rmd:104:331 Moved
The first step of the pipeline is to identify the **Variable Methylated Loci**(VML) in the data set. You might be wondering *"What is a VML and why do we use them instead of DNAme levels from each CpG site?"*. We use **loci** because it is well established that nearby CpG sites are [very likely to share a similar DNAme profile](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6093082/) and therefore work as functional units. Then, from a statistical point of view, testing separately proximal CpGs that are part of the same unit is redundant. On the other side, we use only **variable** regions because we are interested in the units that display a high level of variability; in other words, in non-variant sites there is no variability left to be explained by genetics or environment. So, in conclusion, we use **VML** to increase our power and reduce the multiple hypothesis testing burden by grouping probes that are likely to work as a biological unit, and by only focusing in the set of regions that are of interest of this study.
                                                                                                                                                                                                                                                                                                                                          ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
                                                                                                                                                                                                                                                                                                                                          https://pmc.ncbi.nlm.nih.gov/articles/PMC6093082/
```

---
