### 2026-02-17 f1a4a63
Added rOpenSci badge
