+++ README.md
@@ -12,6 +12,8 @@ stable](https://img.shields.io/badge/lifecycle-stable-brightgreen.svg)](https://
 [![Codecov test
 coverage](https://codecov.io/gh/ErickNavarroD/RAMEN/graph/badge.svg)](https://app.codecov.io/gh/ErickNavarroD/RAMEN)
 [![R-CMD-check](https://github.com/ErickNavarroD/RAMEN/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/ErickNavarroD/RAMEN/actions/workflows/R-CMD-check.yaml)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/743_status.svg)](https://github.com/ropensci/software-review/issues/743)
 <!-- badges: end -->
 
 ## Overview
