+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.2.0
 pkgdown_sha: ~
 articles:
   RAMEN: RAMEN.html
-last_built: 2026-01-13T08:44Z
+last_built: 2026-02-17T21:34Z
 urls:
   reference: https://ericknavarrod.github.io/RAMEN/reference
   article: https://ericknavarrod.github.io/RAMEN/articles
