+++ README.md
@@ -19,12 +19,27 @@ Review](https://badges.ropensci.org/743_status.svg)](https://github.com/ropensci
 ## Overview
 
 Regional Association of Methylome variability with the Exposome and
-geNome (RAMEN) is an R package whose goal is to identify genome-wide
-Variable Methylated Loci (VML) from microarray DNA methylation data.
-Then, using genomic and exposomic data, it can identify which model out
-of the following explains best the DNA methylation variability at each
-VML: genetic (G), environmental (E), additive (G+E) or interactive
-(GxE).
+geNome (RAMEN) is an R package which goal is to estimate the
+contribution of genetic variants and environmental exposures to DNA
+methylation variability at a genome-wide scale. RAMEN takes advantage of
+the fact that DNA methylation levels at nearby CpG sites are often
+correlated, and uses this information to identify Variable Methylated
+Loci (VML) from microarray DNA methylation data. Then, using genomic and
+exposomic data, it can identify which model out of the following
+explains best the DNA methylation variability at each VML: genetic (G),
+environmental (E), additive (G+E) or interactive (GxE).
+
+RAMEN provides a Findable, Accesible, Interoperable and Reusable (FAIR)
+workflow to conduct gene-environment contribution analyses to
+high-dimensional DNA methylome data (described in [Navarro-Delgado et
+al. (2025)](https://doi.org/10.1186/s13059-025-03864-4). Using a blend
+of traditional statistical methods and machine learning approaches,
+RAMEN is designed to be computationally efficient and user-friendly,
+allowing researchers to gain insights into the complex interplay between
+genetics, environment and DNA methylation variability. The package
+includes a detailed tutorial, and individual functions that could be
+useful for other applications beyond the gene-environment contribution
+analysis.
 
 ## Installation
 
@@ -36,38 +51,266 @@ You can install the development version of RAMEN from
 devtools::install_github("ErickNavarroD/RAMEN")
 ```
 
-## Core functions
-
-RAMEN consists of six main functions:
-
-- `findVML()` identifies Variable Methylated Loci (VML) from microarray
-  data
-- `summarizeVML()`summarizes the regional methylation state of each VML
-- `findCisSNPs()` identifies the SNPs in *cis* of each VML
-- `selectVariables()` conducts a LASSO-based feature selection strategy
-  to identify potentially relevant *cis* SNPs and environmental
-  variables
-- `lmGE()` fits linear single-variable genetic (G), environmental (E),
-  pairwise additive (G+E) and pairwise interaction (GxE) linear models,
-  and select the best explanatory model for each VML.
-- `nullDistGE()` simulates a null distribution of G and E effects on
-  DNAme variability. Useful for filtering out poor-performing best
-  explanatory models selected by *lmGE()*.
-
-Altogether, these functions create a pipeline that takes a set of
-individuals with genome, exposome and DNA methylome information, and
-generates an estimation of the contribution of genetic variants and
-environmental exposures to its DNA methylation variability. Functions
-that conduct computationally intensive tasks are compatible with
-parallel computing.
-
-<img src="man/figures/RAMEN_pipeline.png" width="600"/>
-
 ## Usage
 
[…]
