### 2026-03-03 b918dc7
Add comments to examples
