+++ DESCRIPTION
@@ -1,23 +1,20 @@
 Package: RAMEN
-Title: RAMEN: Regional Association of Methylome variability with the Exposome 
-and geNome
+Title: RAMEN: Regional Association of Methylome variability with the Exposome and geNome
 Version: 2.0.0
-Authors@R: 
+Authors@R: c(
     person("Erick I.", "Navarro-Delgado", 
-        email = "ericknadel98@hotmail.com", 
-        role = c("aut", "cre"),
-        comment = c(ORCID = "0000-0003-1040-3519")),
+            email = "ericknadel98@hotmail.com", 
+            role = c("aut", "cre"),
+            comment = c(ORCID = "0000-0003-1040-3519")),
     person("Keegan", "Korthauer", 
-        role = c("aut"), 
-        email = "keegan@stat.ubc.ca", 
-        comment = c(ORCID = "0000-0002-4565-1654")),
+            role = "aut", 
+            email = "keegan@stat.ubc.ca", 
+            comment = c(ORCID = "0000-0002-4565-1654")),
     person("Michael S.", "Kobor", 
-        email = "michael.kobor@ubc.ca",
-        role = "aut"),
-        comment = c(ORCID = "0000-0003-4140-1743"))
-Description: R package that identifies which genetic (G), environmental (E), 
-additive (G+E) or interaction (GxE) effect better explains DNA methylation 
-levels in Variable Methylated Loci using microarray data.
+            email = "michael.kobor@ubc.ca",
+            role = "aut",
+            comment = c(ORCID = "0000-0003-4140-1743")))
+Description: R package that identifies which genetic (G), environmental (E), additive (G+E) or interaction (GxE) effect better explains DNA methylation levels in Variable Methylated Loci using microarray data.
 License: GPL (>= 3)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
