### 2026-03-03 1e000b7
Update acknowledgment section
