### 2026-03-03 3933687
Changed the illumina manifest to Suggest to reduce the number of imported packages.
