+++ DESCRIPTION
@@ -23,6 +23,9 @@ Suggests:
     BiocStyle,
     knitr,
     rmarkdown,
+    IlluminaHumanMethylation450kanno.ilmn12.hg19,
+    IlluminaHumanMethylationEPICanno.ilm10b4.hg19,
+    IlluminaHumanMethylationEPICv2anno.20a1.hg38,
     ggplot2,
     tidyr,
     testthat (>= 3.0.0)
@@ -33,9 +36,6 @@ Imports:
     foreach,
     GenomicRanges,
     glmnet,
-    IlluminaHumanMethylation450kanno.ilmn12.hg19,
-    IlluminaHumanMethylationEPICanno.ilm10b4.hg19,
-    IlluminaHumanMethylationEPICv2anno.20a1.hg38,
     IRanges,
     iterators,
     lifecycle,
