### 2026-03-03 da1ba43
Add packages to the README that are needed for the installation of RAMEN
