+++ README.Rmd
@@ -49,11 +49,16 @@ applications beyond the gene-environment contribution analysis.
 
 ## Installation
 
-You can install the development version of RAMEN from 
+You can install the latest version of RAMEN from 
 [GitHub](https://github.com/) with:
 
 ``` r
 # install.packages("devtools")
+# install.packages("BiocManager")
+# BiocManager::install("S4Vectors")
+# BiocManager::install("IRanges")
+# BiocManager::install("GenomicRanges")
+
 devtools::install_github("ErickNavarroD/RAMEN")
 ```
 
