+++ vignettes/RAMEN.Rmd
@@ -101,7 +101,7 @@ library(tidyr)
 
 ## Identify VML and summarize their methylation state
 
-The first step of the pipeline is to identify the **Variable Methylated Loci**(VML) in the data set. You might be wondering *"What is a VML and why do we use them instead of DNAme levels from each CpG site?"*. We use **loci** because it is well established that nearby CpG sites are [very likely to share a similar DNAme profile](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6093082/) and therefore work as functional units. Then, from a statistical point of view, testing separately proximal CpGs that are part of the same unit is redundant. On the other side, we use only **variable** regions because we are interested in the units that display a high level of variability; in other words, in non-variant sites there is no variability left to be explained by genetics or environment. So, in conclusion, we use **VML** to increase our power and reduce the multiple hypothesis testing burden by grouping probes that are likely to work as a biological unit, and by only focusing in the set of regions that are of interest of this study.
+The first step of the pipeline is to identify the **Variable Methylated Loci**(VML) in the data set. You might be wondering *"What is a VML and why do we use them instead of DNAme levels from each CpG site?"*. We use **loci** because it is well established that nearby CpG sites are [very likely to share a similar DNAme profile](https://doi.org/10.1016/j.stemcr.2018.07.003) and therefore work as functional units. Then, from a statistical point of view, testing separately proximal CpGs that are part of the same unit is redundant. On the other side, we use only **variable** regions because we are interested in the units that display a high level of variability; in other words, in non-variant sites there is no variability left to be explained by genetics or environment. So, in conclusion, we use **VML** to increase our power and reduce the multiple hypothesis testing burden by grouping probes that are likely to work as a biological unit, and by only focusing in the set of regions that are of interest of this study.
 
 RAMEN identifies 2 categories of VML: 
 
