### 2026-03-03 85e13d9
Add a check for R 4.2
