### 2026-03-03 b804f4e
Improved code style - used styler and addressed standardized the use of the native pipe operator
