+++ tests/testthat/test-workflow.R
@@ -112,14 +112,13 @@ test_that("findVML throws errors when expected", {
     ),
     "The method must be either 'mad' or 'variance'. Please select one of those options"
   )
-
 })
 
 test_that("findVML works with EPICv2 probes", {
   epic2_methylation_data <- RAMEN::test_methylation_data
   rownames(epic2_methylation_data) <- data.frame(IlluminaHumanMethylationEPICv2anno.20a1.hg38::Locations) |>
     dplyr::filter(chr == "chr21") |>
-    arrange(chr, pos) |> #Make sure to extract neighbouring probes to have VML
+    arrange(chr, pos) |> # Make sure to extract neighbouring probes to have VML
     slice_head(n = nrow(RAMEN::test_methylation_data)) |>
     rownames()
 
@@ -134,7 +133,7 @@ test_that("findVML works with EPICv2 probes", {
   )
   expect_true(is.list(VML_epic2))
   expect_true(is.data.frame(VML_epic2$VML))
-  expect_equal(ncol(VML_epic2$VML), 10 )
+  expect_equal(ncol(VML_epic2$VML), 10)
 })
 
 test_that("findVML works with var_distribution = 'all' and mad score", {
@@ -149,7 +148,7 @@ test_that("findVML works with var_distribution = 'all' and mad score", {
   )
   expect_true(is.list(VML_allvar))
   expect_true(is.data.frame(VML_allvar$VML))
-  expect_equal(ncol(VML_allvar$VML), 10 )
+  expect_equal(ncol(VML_allvar$VML), 10)
 })
 
 test_that("sVMPs have no correlation", {
@@ -193,11 +192,10 @@ test_that("summarizeVML adds VML_index when not present", {
   )
   expect_true(is.data.frame(summarized_no_index))
   expect_true(all(
-    colnames(summarized_no_index) %in% paste0("VML", seq_len(nrow(VML_no_index))))
-  )
+    colnames(summarized_no_index) %in% paste0("VML", seq_len(nrow(VML_no_index)))
+  ))
   expect_equal(nrow(summarized_no_index), ncol(test_methylation_data))
-}
-)
+})
 
 test_that("summarizeVML values are correct", {
   # First for sVMPs: the summarized value should be equal to the methylation
@@ -270,7 +268,8 @@ test_that("findCisSNPs adds a VML index when it is not present", {
     VML_df = VML$VML |>
       dplyr::select(-VML_index),
     genotype_information = RAMEN::test_genotype_information,
-    distance = 1e+06)
+    distance = 1e+06
+  )
   expect_true("VML_index" %in% colnames(VML_cis_snps_noID))
 })
 
@@ -286,38 +285,43 @@ test_that("findCisSNPs output structure is correct", {
 
 test_that("findCisSNPs throws errors when expected", {
   expect_error(
-    RAMEN::findCisSNPs(VML_df = VML$VML |>
-                         dplyr::select(-seqnames),
-                       genotype_information = RAMEN::test_genotype_information,
-                       distance = 1e+06
-                       ),
+    RAMEN::findCisSNPs(
+      VML_df = VML$VML |>
+        dplyr::select(-seqnames),
+      genotype_information = RAMEN::test_genotype_information,
+      distance = 1e+06
+    ),
     "Please make sure the VML_df object has the required columns with the appropiate names (check documentation for further information)",
     fixed = TRUE
   )
   expect_error(
-    RAMEN::findCisSNPs(VML_df = VML$VML,
-                       genotype_information = RAMEN::test_genotype_information |>
-                         dplyr::select(-CHROM),
-                       distance = 1e+06
+    RAMEN::findCisSNPs(
+      VML_df = VML$VML,
[…]
