+++ data-raw/test_environmental_matrix.R
@@ -7,10 +7,12 @@ test_environmental_matrix <- matrix(rnorm(100 * sample_size, 0, 1),
   ncol = 100
 )
 rownames(test_environmental_matrix) <- paste("ID",
-                                             as.character(1:sample_size),
-                                             sep = "")
+  as.character(1:sample_size),
+  sep = ""
+)
 colnames(test_environmental_matrix) <- paste("E",
-                                             as.character(1:100),
-                                             sep = "")
+  as.character(1:100),
+  sep = ""
+)
 
 usethis::use_data(test_environmental_matrix, overwrite = TRUE)

[…]
