### 2026-03-04 432b425
Separate tests into self-contained units for each function
