+++ tests/testthat/test-workflow.R
@@ -1,719 +1,0 @@
-# Since the input of some functions is the output of others in the package,
-# the whole workflow is going to be tested in this script to minimize the
-# run time.
-
-library(testthat)
-library(dplyr)
-
-#### findVML() ####
-VML <- RAMEN::findVML(
-  methylation_data = RAMEN::test_methylation_data,
-  array_manifest = "IlluminaHumanMethylationEPICv1",
-  cor_threshold = 0,
-  var_method = "variance",
-  var_distribution = "ultrastable",
-  var_threshold_percentile = 0.99,
-  max_distance = 1000
-)
-
-test_that("findVML variance calculation is correct", {
-  probe_test <- VML$highly_variable_probes[1:10, ]
-  observed_variance <- RAMEN::test_methylation_data[probe_test$TargetID, ] |>
-    apply(1, var)
-  names(observed_variance) <- NULL
-  expect_equal(observed_variance, probe_test$var_score)
-})
-
-test_that("findVML output structure is correct", {
-  expect_true(is.list(VML))
-  expect_true("VML" %in% names(VML))
-  expect_true("highly_variable_probes" %in% names(VML))
-  expect_true("var_score_threshold" %in% names(VML))
-  expect_true(is.data.frame(VML$VML))
-  expect_true(is.data.frame(VML$highly_variable_probes))
-})
-
-test_that("findVML handles a different var_method option", {
-  VML_result_mad <- RAMEN::findVML(
-    methylation_data = RAMEN::test_methylation_data,
-    array_manifest = "IlluminaHumanMethylationEPICv1",
-    cor_threshold = 0,
-    var_method = "mad",
-    var_distribution = "ultrastable",
-    var_threshold_percentile = 0.99,
-    max_distance = 1000
-  )
-  probe_test <- VML_result_mad$highly_variable_probes[1:10, ]
-  observed_mad <- RAMEN::test_methylation_data[probe_test$TargetID, ] |>
-    apply(1, mad)
-  names(observed_mad) <- NULL
-  expect_equal(observed_mad, probe_test$var_score)
-  expect_true(is.list(VML_result_mad))
-  expect_true("VML" %in% names(VML_result_mad))
-})
-
-test_that("findVML throws errors when expected", {
-  expect_error(
-    RAMEN::findVML(
-      methylation_data = RAMEN::test_methylation_data,
-      array_manifest = "IlluminaHumanMethylationEPICv1",
-      cor_threshold = 1000,
-      var_method = "ultrastable",
-      var_distribution = "ultrastable",
-      var_threshold_percentile = 0.99,
-      max_distance = 1000
-    ),
-    "'cor_threshold' must be of type 'numeric' and from 0 to 1"
-  )
-  expect_error(
-    RAMEN::findVML(
-      methylation_data = RAMEN::test_methylation_data,
-      array_manifest = "IlluminaHumanMethylationEPICv1",
-      cor_threshold = 0,
-      var_method = "variance",
-      var_distribution = "binomial",
-      var_threshold_percentile = 0.99,
-      max_distance = 1000
-    ),
-    "'var_distribution' must be one of 'all' or 'ultrastable'"
-  )
-  expect_error(
-    RAMEN::findVML(
-      methylation_data = as.matrix(RAMEN::test_methylation_data),
-      array_manifest = "IlluminaHumanMethylationEPICv1",
-      cor_threshold = 0,
-      var_method = "variance",
-      var_distribution = "ultrastable",
-      var_threshold_percentile = 0.99,
-      max_distance = 1000
-    ),
[…]

[…]
