+++ R/zzz.R
@@ -14,11 +14,10 @@
                 '.           .'
                   '--.....--'
 
-If you use RAMEN for your analysis, please cite: Navarro-Delgado, E.I.,
-Czamara, D., Edwards, K. et al. RAMEN: Dissecting individual, additive
-and interactive gene-environment contributions to DNA methylome variability
-in cord blood. Genome Biol 26, 421 (2025).
-https://doi.org/10.1186/s13059-025-03864-4",
+If you use RAMEN for your analysis, please cite: Navarro-Delgado,
+E.I., et al. RAMEN: Dissecting individual, additive and interactive
+gene-environment contributions to DNA methylome variability in cord
+blood. Genome Biol 26, 421 (2025).",
     domain = NULL, appendLF = TRUE
   )
 } # ASCII letters were generated by https://ascii.co.uk/text; the bowl was taken from http://www.geocities.ws/SoHo/7373/food.html and modified
