### 2026-03-04 #743 @ErickNavarroD
Hi @maurolepore
Thank you for all your comments and suggestions! I appreciate the time you spent on this. Here are my responses to your comments. Please let me know if you have any follow-up questions or suggestions.

> ml03. In README I see what RAMEN does. Please also add why it's useful, particularly in comparison with other related packages (see also fit and overlap below). The discussion about the category in which this package fits may help frame its usefulness.

I have added more details in the Overview section, which now has a new paragraph describing why it is useful. I framed the package as a FAIR workflow to conduct gene-environment contribution analyses. Please let me know if more details would be preferred.
Regarding other packages that do similar things, this package is to my knowledge the first one designed to conduct this analysis in the epigenetics field. The methods take inspiration from previous research (cited and acknowledged in the [RAMEN paper]( https://link.springer.com/article/10.1186/s13059-025-03864-4), but all of these previous papers provide only methodological description – they did not develop a package or provide code. Also, RAMEN uses general packages developed to manipulate data (like tidyverse, or GenomicRanges) or fit models (like glmnet), but I am not sure how informative it would be to compare RAMEN with those packages.

> ml04. In README I see the Usage section points to a rich vignette. Also please add a basic example to help understand some representative inputs and outputs as quickly as possible.

I have added examples to the README file as well.

> ml05. In README I see the section "How to get help for RAMEN" mentions GitHub issues but provides no link. Consider adding the link or relying on other standard places (e.g. the website) that automatically include that link.

Thanks for the observation! I added a link that directs the reader to a new issue.

> ml06. In README I see authors not listed in DESCRIPTION. Please review for consistency.

I updated the list of authors that contributed to developing the package in the DESCRIPTION file.

> ml07. If some grouping makes sense, consider using it over the default alphabetical order. I think this is optional but may become more important if the number of functions grows.

I could make a grouping by functions and data sets, but I don’t know how to make these groupings on the webpage, since the Index in the webpage is automatically deployed when I push changes through a GitHub action. Would I have to modify this manually?

> ml08. Please acknowledge the related packages and explain how this package is still useful/necessary.

The answer to this is the same as ml03.

> ml09. Please try the installation instructions in a fresh environment to identify requirements beyond what's documented in README, e.g. Bioconductor packages.

I installed the package in a fresh environment, and added the instructions on how to install the necessary Bioconductor packages

> ml10. Please review .github/CONTRIBUTING.md. I see the placeholder "our_package" and the phrase "this and other tidyverse packages".

Thanks for the catch! I have replaced all the placeholders, and I have reworded the “this and other tidyverse packages” section.

> ml11. On the website I see some functions and datasets do not print the output. Please show a bit of output, e.g. with str() or head(), to help readers understand what to expect even before they run the code themselves. See ([building.html#general](https://devguide.ropensci.org/building.html#general)): All exported package functions should be fully documented with examples.

I added code chunks in the vignette to print the head of all input datasets that are used, and made sure all of the outputs are explored with str() or head().

> ml12. In DESCRIPTION I see an email address that is not the single one used across the package documentation. Please review for consistency.

I have updated the email across the package to make sure the same one is listed.

> ml13. Please add .Rbuildignore to avoid building the tarball with irrelevant files and directories, e.g. data-raw/. You may do this automatically with usethis::use_data_raw()

I already have an .Rbuildignore file, which lists ^data-raw$. Is there anything I need to change in this regard?

> ml14. Consider removing unused .png files if any, and removing docs/ if the website is built on GitHub Actions.

I removed the unused png files and the docs folder

> ml15. Please double check the citation follows the ([dev guide requirements](https://devguide.ropensci.org/building.html#citation-file))

I updated the citation file and made sure it follows the guidelines.

> ml16. DESCRIPTION declares Depends: R (>= 4.2.0) but CI (R-CMD-check.yaml) only tests release, devel, oldrel-1 (~R 4.4). Please consider extending the CI workflow to cover the stated R-version dependency.

I added a check for R 4.2.0 in the ubuntu-latest OS.

> ml17. Consider structuring R/ and tests/ files so that they play well with usethis::use_r() and usethis::use_test(). Beyond the convenience, using a standard structure helps reviewers and contributors navigate an unfamiliar package.

I re-structured the test files, so now each function has its own test-function.R file.

> ml18. Consider rewriting tests so that they are [self-contained tests](https://r-pkgs.org/testing-design.html#sec-testing-design-self-contained). Ideally separate tests should not share state. Tests are easier to understand, debug, and maintain when they avoid code outside test_that(), i.e. no calls to library() (but see withr::with_namespace() and withr::with_package()) nor data shared across tests. See also [testthat special files](https://testthat.r-lib.org/articles/special-files.html), the [fixtures vignette](https://testthat.r-lib.org/articles/test-fixtures.html), the [R Packages testing design](https://r-pkgs.org/testing-design.html#sec-testing-design-principles), and [DRY, DAMP, or WET?](https://ropensci.org/blog/2024/03/18/dry-damp/).

These resources were super helpful! Thank you for sharing them. I struggled when trying to create self-contained tests because my package is a workflow – so the output of one function is usually the input of the next one. I have now added a helper file, there is no code outside test_that() chunks, and there is a test file for each function that is independent of the rest.

> ml19. Please review the test runtime and explore ways to make tests run in a few seconds rather than a few minutes. The faster the tests the more often we run them and the more bugs we catch :-)

I removed a couple of redundant tests and have done modifications as mentioned above.  The time of the tests now is ~25 seconds in my computer. The tests that take the most time are the ones involving the nullDistGE() function, which conducts permutations. I tried reducing the time of those tests as much as I could while still keeping them informative, but this is a time-consuming function.

> ml20. Please review tests so that the test-report prints cleanly. Busy output can hide important details and confuse reviewers.

I suppressed the progress messages from running the functions and the test report is now clean.

> ml21. I see both %>% and |>. Consider using one type of pipe consistently.

I updated the files and now use |> across all functions.

> ml22. Consider finding opportunities to deduplicate code (e.g. c(list(NULL), list("")) and simplify logic.

I will keep this in mind as I modify the package in future instances. I have modified the lines of code you are referring to with c(list(NULL),…), and have simplified them.

> ml23. In map_revmap_names consider removing \dontrun{} since it seems unnecessary.

I had issues with this function because it is only internal (not exported). So, initially it did not have examples. When running pkgcheck(), the function flagged that the map_rev_names() function did not have examples. But when I added an example, the build check would fail because the function is not exported, so the code in the example threw an error. So, the only way I found to pass the pkgcheck checks and not have an error was to use \dontrun{}.

> ml24. In the output of pkgcheck I see "Only 1 non-exported function vs 8 exported. High cyclocomplexity + low non-exported function count suggests core functions are large monoliths that could be broken into smaller internal helpers."

I will keep this in mind for future modifications of the package – encapsulate small processes in functions that are used in the main functions. Thanks!

> ml25. Please style the code with an automated tool like air format . for readability. Using a standard code style helps reviewers and contributors more easily understand the code.

I used styler (style_pkg()) to style the code.

> ml26. Please address the results of urlchecker::url_check()

I addressed the warning produced by url_check()

Also, I added the software review badge in the README.md as requested by the bot.

Thank you again for all your feedback!
