### 2026-04-03 #743 @maurolepore
Still no luck. I sent 10 invitations, 3 declined, 7 waiting. It's unfortunate but not uncommon.
