### 2026-04-23 cfd3d59
Replaced install_github with pak
