+++ README.Rmd
@@ -59,7 +59,7 @@ You can install the latest version of RAMEN from
 # BiocManager::install("IRanges")
 # BiocManager::install("GenomicRanges")
 
-devtools::install_github("ErickNavarroD/RAMEN")
+pak::pak("ErickNavarroD/RAMEN")
 ```
 
 ## Usage
