### 2026-05-03 #743 @maurolepore
Just a short note for the record: my new EiC rotation has begun. I'll continue here as the handling editor.

Should anyone need support from someone else, feel free to ping direclty our software-review lead @mpadge.
