### 2026-05-10 #743 @maurolepore
Thanks @llrs for your review!

@ErickNavarroD please wait for the second review to land. I can't find the policy in our guide but we prefer you to respond once we have the more complete picture that both reviews provide.

@maelle do you remember if we have a sentence somewhere about this that I could quote and link?
