### 2026-05-16 #743 @maurolepore
@vafadaru thank you so much for your review 🙏

Could you please copy paste this into a new comment and complete this piece of the reviewer template? I particularly need the estimated time in hours so I can invoke the bot command that logs that in our database AND stops bugging you about due dates 😄

```
Estimated hours spent reviewing:

```
