+++ NEWS.md
@@ -1,3 +1,5 @@
+# RAMEN (development version)
+
 # RAMEN 2.0.1 - June 30, 2026
 
 This minor version presents no changes from the user perspective. Code 
