+++ reference/summarizeVML.md
@@ -45,14 +45,14 @@ the function can be run as usual.
 ``` r
 ## Find VML in test data
 VML <- RAMEN::findVML(
-   methylation_data = RAMEN::test_methylation_data,
-   array_manifest = "IlluminaHumanMethylationEPICv1",
-   cor_threshold = 0,
-   var_method = "variance",
-   var_distribution = "ultrastable",
-   var_threshold_percentile = 0.99,
-   max_distance = 1000
-   )
+  methylation_data = RAMEN::test_methylation_data,
+  array_manifest = "IlluminaHumanMethylationEPICv1",
+  cor_threshold = 0,
+  var_method = "variance",
+  var_distribution = "ultrastable",
+  var_threshold_percentile = 0.99,
+  max_distance = 1000
+)
 #> Identifying Highly Variable Probes...
 #> Identifying sparse Variable Methylated Probes
 #> Identifying Variable Methylated Regions...
@@ -62,5 +62,5 @@ VML <- RAMEN::findVML(
 summarized_VML <- RAMEN::summarizeVML(
   VML_df = VML$VML,
   methylation_data = RAMEN::test_methylation_data
-  )
+)
 ```
