+++ reference/nullDistGE.md
@@ -139,22 +139,22 @@ pool the delta R squared values from all VML to create a null
 distribution taking advantage of the high number of VML in the dataset.
 This assumption decreases significantly the number of permutations
 required to create a null distribution and reduces the computational
-time. For further information please read the RAMEN paper (in
-preparation).
+time. For further information please read the RAMEN paper
+(https://doi.org/10.1186/s13059-025-03864-4).
 
 ## Examples
 
 ``` r
 ## Find VML in test data
 VML <- RAMEN::findVML(
-   methylation_data = RAMEN::test_methylation_data,
-   array_manifest = "IlluminaHumanMethylationEPICv1",
-   cor_threshold = 0,
-   var_method = "variance",
-   var_distribution = "ultrastable",
-   var_threshold_percentile = 0.99,
-   max_distance = 1000
-   )
+  methylation_data = RAMEN::test_methylation_data,
+  array_manifest = "IlluminaHumanMethylationEPICv1",
+  cor_threshold = 0,
+  var_method = "variance",
+  var_distribution = "ultrastable",
+  var_threshold_percentile = 0.99,
+  max_distance = 1000
+)
 #> Identifying Highly Variable Probes...
 #> Identifying sparse Variable Methylated Probes
 #> Identifying Variable Methylated Regions...
@@ -164,26 +164,26 @@ VML_with_cis_snps <- RAMEN::findCisSNPs(
   VML_df = VML$VML,
   genotype_information = RAMEN::test_genotype_information,
   distance = 1e6
-  )
+)
 #> Reminder: please make sure that the positions of the VML data frame and the ones in the genotype information are from the same genome build.
 
 ## Summarize methylation levels in VML
 summarized_methyl_VML <- RAMEN::summarizeVML(
- methylation_data = RAMEN::test_methylation_data,
- VML_df = VML_with_cis_snps
- )
+  methylation_data = RAMEN::test_methylation_data,
+  VML_df = VML_with_cis_snps
+)
 
 ## Simulate null distribution of G and E contributions on DNAme variability
 null_dist <- RAMEN::nullDistGE(
-   VML_df = VML_with_cis_snps,
-   genotype_matrix = RAMEN::test_genotype_matrix,
-   environmental_matrix = RAMEN::test_environmental_matrix,
-   summarized_methyl_VML = summarized_methyl_VML,
-   permutations = 5,
-   covariates = RAMEN::test_covariates,
-   seed = 1,
-   model_selection = "AIC"
- )
+  VML_df = VML_with_cis_snps,
+  genotype_matrix = RAMEN::test_genotype_matrix,
+  environmental_matrix = RAMEN::test_environmental_matrix,
+  summarized_methyl_VML = summarized_methyl_VML,
+  permutations = 5,
+  covariates = RAMEN::test_covariates,
+  seed = 1,
+  model_selection = "AIC"
+)
 #> Starting permutation 1 of 5
 #> Starting variable selection of permutation 1 of 5
 #> Starting lmGE in permutation 1 of 5
