+++ reference/medCorVMR.md
@@ -43,18 +43,21 @@ options(future.globals.maxSize= +Inf).
 ## Examples
 
 ``` r
-#Create a VML data.frame
-VMR_df <- data.frame(seqnames = c("chr21", "chr21"),
+
+# Create a VML data.frame
+VMR_df <- data.frame(
+  seqnames = c("chr21", "chr21"),
   start = c(10861376, 10862171),
   end = c(10862507, 10883548),
-  probes =  I(list(c("cg15043638", "cg18287590", "cg17975851"),
-                   c("cg13893907", "cg17035109", "cg06187584"))))
+  probes = I(list(
+    c("cg15043638", "cg18287590", "cg17975851"),
+    c("cg13893907", "cg17035109", "cg06187584")
+  ))
+)
 
 # Compute median correlation for each VMR
 medCorVMR(VMR_df = VMR_df, methylation_data = RAMEN::test_methylation_data)
 #>   seqnames    start      end       probes median_correlation
 #> 1    chr21 10861376 10862507 cg150436....          0.7275160
 #> 2    chr21 10862171 10883548 cg138939....          0.7468144
-
-
 ```

[…]
