### 2026-06-30 981a1a3
Mention why the goal of the package is worth pursuing and who should be using the package
