+++ vignettes/RAMEN.Rmd
@@ -27,7 +27,30 @@ knitr::opts_chunk$set(
 ```
 # Introduction 
 
-**Regional Association of Methylome variability with the Exposome and geNome (RAMEN)** is an R package whose goal is to integrate genomic, methylomic and exposomic data to model the contribution of genetics (G) and the environment (E) to DNA methylation (DNAme) variability. RAMEN identifies Variable Methylated Loci (VML) in microarray DNAme data and then, using genotype and environmental data, it identifies which of the following models better explains this variability in regions across the methylome: 
+**Regional Association of Methylome variability with the Exposome and geNome**
+**(RAMEN)** is an R package which goal is to estimate the contribution of genetic 
+variants and environmental exposures to loci with high DNA methylation (DNAme) 
+variability at a genome-wide scale using population data. Characterizing the 
+factors that contribute to DNAme variability is important because DNAme is a key
+epigenetic mechanism that regulates gene expression and plays an important role 
+in development, disease, and environmental adaptation.
+
+RAMEN provides a Findable, Accesible, Interoperable and Reusable (FAIR) workflow
+to conduct gene-environment contribution analyses to high-dimensional DNA 
+methylome data (described in 
+[Navarro-Delgado et al. (2025)](https://doi.org/10.1186/s13059-025-03864-4). 
+Using a blend of traditional statistical methods and machine learning 
+approaches, RAMEN is designed to be computationally efficient and user-friendly,
+allowing researchers to gain insights into the complex interplay between 
+genetics, environment and DNA methylation variability. The package includes a 
+detailed tutorial, and individual functions that could be useful for other 
+applications beyond the gene-environment contribution analysis.
+
+RAMEN takes advantage of the fact that DNA methylation levels
+at nearby CpG sites are often correlated, and uses this information to identify 
+Variable Methylated Loci (VML) from microarray DNA methylation data. Then, 
+integrating genomic and exposomic data, it can identify which model out of the 
+following explains best the DNA methylation variability at each VML:
 
 ```{r modelstable, echo=FALSE}
 library(knitr)
@@ -41,26 +64,44 @@ models <- data.frame(
 kable(models, caption = "Fitted models")
 ```
 
-where G variables are represented by SNPs, E variables by environmental exposures, and where covars are concomitant variables (i.e. variables that are adjusted for in the model and not of interest in the study such as cell type proportion, age, etc.).
+where G variables are represented by SNPs, E variables by environmental 
+exposures, and where covars are concomitant variables (i.e. variables that are
+adjusted for in the model and not of interest in the study such as cell type 
+proportion, age, population structure, etc.).
 
-The main [gene-environment interaction modeling][ Gene-environment interaction analysis] pipeline is conducted though six core functions: 
+The main [gene-environment interaction modeling][ Gene-environment interaction analysis] 
+pipeline is conducted though six core functions: 
 
 -   `findVML()` identifies Variable Methylated Regions (VML) in microarrays
 -   `summarizeVML()`summarizes the regional methylation state of each VML
 -   `findCisSNPs()` identifies the SNPs in *cis* of each VML
--   `selectVariables()` conducts a LASSO-based variable selection strategy to identify potentially relevant *cis* SNPs and environmental variables
--   `lmGE()` fits linear single-variable genetic (G) and environmental (E), and pairwise additive (G+E) and interaction (GxE) linear models and select the best explanatory model per VML.
--   `nullDistGE()` simulates a delta R squared null distribution of G and E effects on DNAme variability. Useful for filtering out poor-performing best explanatory models selected by *lmGE()*.
[…]
