### 2026-06-30 c566a62
Replace pak with BiocManager install
