### 2026-07-03 57fa183
Add information about data expectations and general guidance on pre-processing steps.
