+++ README.Rmd
@@ -226,6 +226,12 @@ final_res %>%
   ggplot2::theme_classic()
 ```
 
+It is worth mentioning that RAMEN assumes that all data sets (genome, exposome 
+and methylome) have undergone quality control, pre-processing and normalization 
+steps when required. The choice of methods for these steps are out of the scope 
+of this package, but we provide some resources and guidance in the 
+[tutorial](https://ericknavarrod.github.io/RAMEN/articles/RAMEN.html). 
+
 ## Variations to the standard workflow
 
 Besides using RAMEN for a gene-environment contribution analysis, the package 
