+++ vignettes/RAMEN.Rmd
@@ -105,9 +105,12 @@ If you use RAMEN for any of your analyses, please cite the following publication
 
 # Gene-environment interaction analysis
 
-The main purpose of the RAMEN package is to conduct a methylome-wide analysis to identify which model (G, E, G+E or GxE) better explains the variability across the genome. In this vignette, we will illustrate how to use the package. 
+The main purpose of the RAMEN package is to conduct a methylome-wide analysis to
+identify which model (G, E, G+E or GxE) better explains the variability across 
+the genome. In this vignette, we will illustrate how to use the package. 
 
-To conduct this analysis, the following cleaned data sets (i.e. after quality and exploratory data analysis checks) from a cohort are required: 
+To conduct this analysis, we will use the following data sets from a population 
+cohort:
 
   -   DNAme data
   -   DNAme array manifest
@@ -116,7 +119,134 @@ To conduct this analysis, the following cleaned data sets (i.e. after quality an
   -   Environmental exposure data
   -   Concomitant variables data
   
-Once that we have that data, the overview of the pipeline is the following:
+## Data expectations
+
+RAMEN expects all data sets (genome, exposome, methylome and covariates) to have
+undergone quality control, pre-processing and normalization when required. The 
+choice of methods to do so is beyond the scope of this vignette, but we provide 
+some broad recommendations and resources here. 
+
+### Genomic data 
+
+The QC of genotyping data is a crucial step in any genetic study, as it ensures 
+the accuracy and reliability of the results. We recommend the following steps 
+for QC of microarray genotyping data:
+
+  -   Remove unreliable samples: The specific criteria for removing samples will 
+  vary depending on the genotyping platform, but common criteria include 
+  removing samples with low call rate, chromosome aneuploidy, or indicators of 
+  sample mislabeling such as genotype mismatches when comparing to SNPs 
+  in the DNAme microarray, or reported-predicted sex mismatches in neonatal 
+  samples. If working with Illumina genotyping data, we recommend to see their
+  [technical note](https://www.illumina.com/Documents/products/technotes/technote_infinium_genotyping_data_analysis.pdf)
+  or the RAMEN paper 
+  [Supplementaty Methods](https://link.springer.com/article/10.1186/s13059-025-03864-4#Sec10)
+  for specific cutoffs. 
+  -   Remove unreliable/uninformative SNPs: The specific criteria for removing 
+  SNPs may also vary depending on the genotyping platform, but common criteria 
+  include removing SNPs with low call frequency, genotype clustering metrics, 
+  low minor allele frequencies, or deviations from the Hardy-Weinberg equilibrium.
+  If working with Illumina genotyping data, we recommend to see their
+  [technical note](https://www.illumina.com/Documents/products/technotes/technote_infinium_genotyping_data_analysis.pdf)
+  or the RAMEN paper 
+  [Supplementaty Methods](https://link.springer.com/article/10.1186/s13059-025-03864-4#Sec10)
+  for specific cutoffs and more detailed information.
+  -   Remove related samples: We recommend using PLINK's PI_HAT metric to 
+  identify and remove related samples if working with a genetically-homogeneous 
+  population. Otherwise, we recommend using the 
+  [KING-robust kinship estimator](https://doi.org/10.1093/bioinformatics/btq559)
+  also implemented in [PLINK](https://www.cog-genomics.org/plink/2.0/distance). 
+  -   Remove genetic variants in sex chromosomes.
+  -   (Optional) Impute missing genotypes: To increase the number of genetic 
+  variants available for analysis, we recommend using imputation methods to infer
+  missing genotypes based on the observed genotypes and a reference panel. We 
+  recommend checking out the
[…]
