### 2026-07-03 7dd8a11
Split the text so lines are not too long
