+++ vignettes/RAMEN.Rmd
@@ -254,13 +254,23 @@ knitr::include_graphics("RAMEN_pipeline.png")
 
 where:  
 
-  -   DNAme data is grouped into VML, and then the DNAme state per individual is summarized in each VML. 
-  -   Using the identified VML and the genomic information, we identify the SNPs in *cis* for each VML
-  -   Both the *cis* SNPs and the exposome data are subjected to the variable selection stage
-  -   The selected variables (Single Nucleotide Polymorphisms a and Environmental Exposures) enter the modelling stage, which outputs one single winning model per VML
-  -   The thresholds obtained from the simulated null distribution are used to remove winning models which performance are likely to be due to chance. 
+  -   DNAme data is grouped into VML, and then the DNAme state per individual is
+  summarized in each VML. 
+  -   Using the identified VML and the genomic information, we identify the SNPs
+  in *cis* for each VML
+  -   Both the *cis* SNPs and the exposome data are subjected to the variable 
+  selection stage
+  -   The selected variables (Single Nucleotide Polymorphisms a and Environmental
+  Exposures) enter the modelling stage, which outputs one single winning model 
+  per VML
+  -   The thresholds obtained from the simulated null distribution are used to 
+  remove winning models which performance are likely to be due to chance. 
 
-In the following sections we will go through each of these steps and guide the user regarding the recommended parameters to use in each function of the package. For illustration purposes, we provide small toy data sets that do not intend to simulate the real biological phenomenon. These data sets are already available in the RAMEN package. 
+In the following sections we will go through each of these steps and guide the 
+user regarding the recommended parameters to use in each function of the 
+package. For illustration purposes, we provide small toy data sets that do not 
+intend to simulate the real biological phenomenon. These data sets are already 
+available in the RAMEN package. 
 
 ## Install RAMEN and set up environment
 
@@ -286,20 +296,77 @@ library(tidyr)
 
 ## Identify VML and summarize their methylation state
 
-The first step of the pipeline is to identify the **Variable Methylated Loci**(VML) in the data set. You might be wondering *"What is a VML and why do we use them instead of DNAme levels from each CpG site?"*. We use **loci** because it is well established that nearby CpG sites are [very likely to share a similar DNAme profile](https://doi.org/10.1016/j.stemcr.2018.07.003) and therefore work as functional units. Then, from a statistical point of view, testing separately proximal CpGs that are part of the same unit is redundant. On the other side, we use only **variable** regions because we are interested in the units that display a high level of variability; in other words, in non-variant sites there is no variability left to be explained by genetics or environment. So, in conclusion, we use **VML** to increase our power and reduce the multiple hypothesis testing burden by grouping probes that are likely to work as a biological unit, and by only focusing in the set of regions that are of interest of this study.
+The first step of the pipeline is to identify the **Variable Methylated Loci**
+(VML) in the data set. You might be wondering 
+*"What is a VML and why do we use them instead of DNAme levels from each CpG site?"*
+. We use **loci** because it is well established that nearby CpG sites are 
+[very likely to share a similar DNAme profile](https://doi.org/10.1016/j.stemcr.2018.07.003)
+and therefore work as functional units. Then, from a statistical point of view, 
+testing separately proximal CpGs that are part of the same unit is redundant. 
+On the other side, we use only **variable** regions because we are interested 
+in the units that display a high level of variability; in other words, in 
+non-variant sites there is no variability left to be explained by genetics or 
+environment. So, in conclusion, we use **VML** to increase our power and reduce
+the multiple hypothesis testing burden by grouping probes that are likely to 
+work as a biological unit, and by only focusing in the set of regions that are 
+of interest of this study.
 
 RAMEN identifies 2 categories of VML: 
 
[…]
