### 2026-07-03 d649139
Fix number of cores to 2 in vignette, examples and tests
