+++ R/summarizeVML.R
@@ -17,6 +17,8 @@
 #'
 #' @examples
 #' ## Find VML in test data
+#' # Set the parallel backend to use 2 workers
+#' doParallel::registerDoParallel(2)
 #' VML <- RAMEN::findVML(
 #'   methylation_data = RAMEN::test_methylation_data,
 #'   array_manifest = "IlluminaHumanMethylationEPICv1",
