+++ R/selectVariables.R
@@ -1,28 +1,108 @@
-#' Selection of relevant environment and genotype variables associated with Variably Methylated Loci (VML)
+#' Selection of relevant environment and genotype variables associated with
+#' Variably Methylated Loci (VML)
 #'
-#' For each VML, this function selects potentially relevant genotype and environmental variables associated with DNA methylation levels of said VML using LASSO. See details below for more information.
+#' For each VML, this function selects potentially relevant genotype and
+#'  environmental variables associated with DNA methylation levels of said VML
+#'  using LASSO. See details below for more information.
 #'
-#' selectVariables() uses LASSO, which is an embedded variable selection method that penalizes models that are more complex (i.e., that contain more variables) in favor of simpler models (i.e. that contain less variables), but not at the expense of reducing predictive power. Using LASSO's variable screening property (with high probability, the LASSO estimated model includes the substantial covariates and drops the redundant ones) this function selects genotype and environment variables with potential relevance in the Variable Methylated Loci (VML) dataset (see also Bühlmann and van de Geer, 2011). For each VML, LASSO is run three times: 1) including only the genotype variables for the selection step, 2) including only the environmental variables for the selection step, and 3) Including both the genotype and environmental variables in the selection step. This is done to ensure that the function captures the variables that are relevant within their own category (e.g., SNPs that are strongly associated with the DNAme levels of a VML in the presence of the rest of the SNPs) or in the presence of the variables of the other category (e.g. SNPs that are strongly associated with the DNAme levels of a VML in the presence of the rest of BOTH the SNPs AND environmental variables). Every time LASSO is run, the basal covariates (i.e., concomitant variables )indicated in the argument *covariates* are not penalized (i.e., those variables are always included in the models and their coefficients are not subjected to shrinkage). That way, only the most promising E and G variables in the presence of the concomitant variables will be selected.
+#' selectVariables() uses LASSO, which is an embedded variable selection method
+#' that penalizes models that are more complex (i.e., that contain more
+#' variables) in favor of simpler models (i.e. that contain less variables), but
+#' not at the expense of reducing predictive power. Using LASSO's variable
+#' screening property (with high probability, the LASSO estimated model includes
+#' the substantial covariates and drops the redundant ones) this function
+#' selects genotype and environment variables with potential relevance in the
+#' Variable Methylated Loci (VML) dataset (see also Bühlmann and van de Geer,
+#' 2011). For each VML, LASSO is run three times: 1) including only the genotype
+#' variables for the selection step, 2) including only the environmental
+#' variables for the selection step, and 3) Including both the genotype and
+#' environmental variables in the selection step. This is done to ensure that
+#' the function captures the variables that are relevant within their own
+#' category (e.g., SNPs that are strongly associated with the DNAme levels of a
+#' VML in the presence of the rest of the SNPs) or in the presence of the
+#' variables of the other category (e.g. SNPs that are strongly associated with
+#' the DNAme levels of a VML in the presence of the rest of BOTH the SNPs AND
+#' environmental variables). Every time LASSO is run, the basal covariates
+#' (i.e., concomitant variables )indicated in the argument *covariates* are not
+#' penalized (i.e., those variables are always included in the models and their
+#' coefficients are not subjected to shrinkage). That way, only the most
+#' promising E and G variables in the presence of the concomitant variables
+#' will be selected.
 #'
[…]
