### 2026-07-06 5cde1ef
Shorten running time of examples
