+++ R/findCisSNPs.R
@@ -38,8 +38,9 @@
 #'   max_distance = 1000
 #' )
 #' ## Find cis SNPs around VML
+#' # Use only 5 for demonstration purposes
 #' VML_with_cis_snps <- RAMEN::findCisSNPs(
-#'   VML_df = VML$VML,
+#'   VML_df = VML$VML[1:5, ],
 #'   genotype_information = RAMEN::test_genotype_information,
 #'   distance = 1e6
 #' )
@@ -81,7 +82,7 @@ findCisSNPs <- function(VML_df, genotype_information, distance = 1e6) {
       names = genotype_information$ID
     )
   )
-  # Extend each VML 1 Mb up and downstream
+  #### Extend each VML 1 Mb up and downstream ####
   VML_extended <- VML_gr + distance
 
   VML_df_with_cisSNPs <- VML_df
