+++ man/findCisSNPs.Rd
@@ -52,8 +52,9 @@ VML <- RAMEN::findVML(
   max_distance = 1000
 )
 ## Find cis SNPs around VML
+# Use only 5 for demonstration purposes
 VML_with_cis_snps <- RAMEN::findCisSNPs(
-  VML_df = VML$VML,
+  VML_df = VML$VML[1:5, ],
   genotype_information = RAMEN::test_genotype_information,
   distance = 1e6
 )
