+++ man/nullDistGE.Rd
@@ -120,8 +120,8 @@ please read the RAMEN paper (https://doi.org/10.1186/s13059-025-03864-4).
 # Set the parallel backend to use 2 workers
 doParallel::registerDoParallel(2)
 ## Find VML in test data
-VML <- RAMEN::findVML(
-  methylation_data = RAMEN::test_methylation_data,
+VML <- findVML(
+  methylation_data = test_methylation_data,
   array_manifest = "IlluminaHumanMethylationEPICv1",
   cor_threshold = 0,
   var_method = "variance",
@@ -130,27 +130,29 @@ VML <- RAMEN::findVML(
   max_distance = 1000
 )
 ## Find cis SNPs around VML
-VML_with_cis_snps <- RAMEN::findCisSNPs(
-  VML_df = VML$VML,
-  genotype_information = RAMEN::test_genotype_information,
+VML_with_cis_snps <- findCisSNPs(
+  # Use only 5 for demonstration purposes
+  VML_df = VML$VML[1:5, ],
+  genotype_information = test_genotype_information,
   distance = 1e6
 )
 
 ## Summarize methylation levels in VML
-summarized_methyl_VML <- RAMEN::summarizeVML(
-  methylation_data = RAMEN::test_methylation_data,
+summarized_methyl_VML <- summarizeVML(
+  methylation_data = test_methylation_data,
   VML_df = VML_with_cis_snps
 )
 
 ## Simulate null distribution of G and E contributions on DNAme variability
 ## We will only run one permutation for demonstration purposes
-null_dist <- RAMEN::nullDistGE(
+null_dist <- nullDistGE(
   VML_df = VML_with_cis_snps,
-  genotype_matrix = RAMEN::test_genotype_matrix,
-  environmental_matrix = RAMEN::test_environmental_matrix,
+  genotype_matrix = test_genotype_matrix,
+  environmental_matrix = test_environmental_matrix,
   summarized_methyl_VML = summarized_methyl_VML,
-  permutations = 5,
-  covariates = RAMEN::test_covariates,
+  # Use one permutation for demonstration purposes
+  permutations = 1,
+  covariates = test_covariates,
   seed = 1,
   model_selection = "AIC"
 )
