### 2026-07-08 53dac43
Remove magittr pipe
