+++ NAMESPACE
@@ -1,6 +1,5 @@
 # Generated by roxygen2: do not edit by hand
 
-export("%>%")
 export(findCisSNPs)
 export(findVML)
 export(lmGE)
@@ -12,4 +11,3 @@ importFrom(doRNG,"%dorng%")
 importFrom(foreach,"%do%")
 importFrom(foreach,"%dopar%")
 importFrom(lifecycle,deprecated)
-importFrom(magrittr,"%>%")
