+++ DESCRIPTION
@@ -37,6 +37,7 @@ Suggests:
     IlluminaHumanMethylationEPICv2anno.20a1.hg38,
     ggplot2,
     tidyr,
+    doParallel,
     testthat (>= 3.0.0)
 Config/testthat/edition: 3
 Imports: 
@@ -48,7 +49,6 @@ Imports:
     IRanges,
     iterators,
     lifecycle,
-    magrittr,
     relaimpo,
     S4Vectors,
     stats,
