### 2026-07-08 4b9533b
Create helper functions for checking arguments
