### 2026-07-08 49ebbd4
Specify origin of registerDoParallel
