+++ vignettes/RAMEN.Rmd
@@ -295,7 +295,7 @@ library(tidyr)
 library(doParallel)
 
 # Set the parallel backend to use 2 workers
-registerDoParallel(2)
+doParallel::registerDoParallel(2)
 ```
 
 ## Identify VML and summarize their methylation state
