### 2026-07-08 c8663ee
improved argument checks with helper functions
