+++ R/findVML.R
@@ -38,22 +38,22 @@ map_revmap_names <- function(positions, manifest_hvp) {
 #' Methylated Regions and sparse Variable Methylated Probes. See Details below
 #' for more information.
 #'
-#' This function identifies HVPs based on MAD scores or variance, and groups
-#' them into VML, which are defined as genomic regions with high DNA methylation
-#'  variability.To best capture methylome variability patterns in microarrays,
-#'  we identify two types of VML: Variably Methylated Regions (VMRs) and sparse
-#'   Variably Methylated Probes (sVMPs) .
+#' This function identifies HVPs based on variance or MAD scores, and groups
+#' them into VML, which are defined as genomic locations with high DNA methylation
+#' variability.To best capture methylome variability patterns in microarrays,
+#' we identify two types of VML: Variably Methylated Regions (VMRs) and sparse
+#' Variably Methylated Probes (sVMPs) .
 #'
 #' In one hand, we defined VMRs as two or more proximal highly variable probes
 #' (default: < 1kb apart) with correlated DNAme level (default: r > 0.15).
-#'  Modelling DNAme variability through regions rather than individual CpGs
-#'  provides several methodological advantages in association studies, since
-#'  CpGs display a significant correlation for co-methylation when they are
-#'  close (less than or equal to 1 kilobase). Modelling DNAme variability
-#'  through regions rather than individual CpGs provides several methodological
-#'  advantages in association studies, since CpGs display a significant
-#'  correlation for co-methylation when they are close (less than or equal to 1
-#'   kilobase)
+#' Modelling DNAme variability through regions rather than individual CpGs
+#' provides several methodological advantages in association studies, since
+#' CpGs display a significant correlation for co-methylation when they are
+#' close (less than or equal to 1 kilobase). Modelling DNAme variability
+#' through regions rather than individual CpGs provides several methodological
+#' advantages in association studies, since CpGs display a significant
+#' correlation for co-methylation when they are close (less than or equal to 1
+#' kilobase)
 #'
 #' In addition to traditional VMRs, we also identified sparse Variably
 #' Methylated Probes (sVMPs), a second type of VML that takes into account the
@@ -61,77 +61,74 @@ map_revmap_names <- function(positions, manifest_hvp) {
 #' tailor our analysis to this DNAme platform. sVMPs aimed to retain genomic
 #' regions with high DNAme variability measured by single probes, where probe
 #' grouping based on proximity and correlation is therefore not applicable. This
-#'  is particularly relevant in the Illumina EPIC v1 array, where most covered
-#'  regulatory regions (up to 93%) are represented by just one probe. Notably,
-#'  based on empirical comparisons with whole-genome bisulfite sequencing data,
-#'   these single probes are mostly representative of local regional DNAme
-#'   levels due to their positioning (98.5-99.5%)
+#' is particularly relevant in the Illumina EPIC v1 array, where most covered
+#' regulatory regions (up to 93%) are represented by just one probe. Notably,
+#' based on empirical comparisons with whole-genome bisulfite sequencing data,
+#' these single probes are mostly representative of local regional DNAme
+#' levels due to their positioning (98.5-99.5%)
 #'
 #' This function uses GenomicRanges::reduce() to group the regions, which is
 #' strand-sensitive. In the Illumina microarrays, the MAPINFO for all the probes
-#'  is usually provided for the + strand. If you are using this array, we
-#'  recommend to first convert the strand of all the probes to "+".
+#' is usually provided for the + strand. If you are using this array, we
+#' recommend to first convert the strand of all the probes to "+".
 #'
 #' This function supports parallel computing for increased speed. To do so, you
[…]
