### 2026-07-08 7e62a97
Updated tests after implementing argument checks with helper functions
