### 2026-07-08 38e2e3a
Function now takes a GRanges object instead of a data frame
