+++ R/medCorVMR.R
@@ -12,13 +12,11 @@
 #' function can be run as usual. It is recommended to also set
 #' options(future.globals.maxSize= +Inf).
 #'
-#' @param VMR_df GRanges object converted to a data frame. Must contain the
-#' following columns: "seqnames", "start", "end"  (all of which are produced
-#' automatically when doing the object conversion) and "probes" (containing a
-#' list in which each element contains a vector with the probes constituting the
-#' VMR).
+#' @param VML GRanges object. Must contain a metadata column named "probes",
+#' where each element contains a vector with the probes constituting the
+#' VML.
 #' @inheritParams findVML
-#' @return A data frame like VMR_df with an extra column per region containing
+#' @return A GRanges object like VML with an extra column per region containing
 #' the median pairwise correlation.
 #'
 #' @importFrom foreach %dopar%
@@ -27,27 +25,27 @@
 #' @examples
 #' # Set the parallel backend to use 2 workers
 #' doParallel::registerDoParallel(2)
-#' # Create a VML data.frame
-#' VMR_df <- data.frame(
-#'   seqnames = c("chr21", "chr21"),
-#'   start = c(10861376, 10862171),
-#'   end = c(10862507, 10883548),
-#'   probes = I(list(
-#'     c("cg15043638", "cg18287590", "cg17975851"),
-#'     c("cg13893907", "cg17035109", "cg06187584")
-#'   ))
-#' )
+#' # Create a VML object
+#' VML <- GenomicRanges::GRanges(seqnames = c("chr21", "chr21"),
+#'       ranges = IRanges::IRanges(start = c(10861376, 10862171),
+#'                        end = c(10862507, 10883548)),
+#'       probes = I(list(
+#'         c("cg15043638", "cg18287590", "cg17975851"),
+#'         c("cg13893907", "cg17035109", "cg06187584")))
+#'       )
 #'
 #' # Compute median correlation for each VMR
-#' medCorVMR(VMR_df = VMR_df, methylation_data = RAMEN::test_methylation_data)
+#' medCorVMR(VML = VML, methylation_data = RAMEN::test_methylation_data)
 #'
-medCorVMR <- function(VMR_df, methylation_data) {
-  if (!is.list(VMR_df$probes)) {
-    stop("Please make sure the 'probes' column in VMR_df is a column of lists")
+medCorVMR <- function(VML, methylation_data) {
+  argument_check(VML, "GRanges")
+  if (!"probes" %in% colnames(mcols(VML))) {
+    stop("Please make sure the VML object has the 'probes' column.")
   }
+  argument_check(methylation_data, "data.frame")
   # generate a list where each element will contain a vector with the probes
   # present in one VMR
-  VMR_probes <- VMR_df$probes
+  VMR_probes <- S4Vectors::mcols(VML)$probes
   # Compute correlations
   i <- NULL # Bind variable to the environment
   median_correlation <- foreach::foreach(
@@ -79,6 +77,6 @@ medCorVMR <- function(VMR_df, methylation_data) {
     }
   }
 
-  VMR_df$median_correlation <- median_correlation
-  return(VMR_df)
+  S4Vectors::mcols(VML)$median_correlation <- median_correlation
+  return(VML)
 }
