### 2026-07-08 6240e15
Reduce data casting
