### 2026-07-08 952c909
Improve argument check with helper functions
