### 2026-07-09 42271ee
Adapt this function to work with GRanges objects and reduce data casting
