### 2026-07-09 fafa609
Removed unnecesary check
