### 2026-07-09 69ada79
Updated tests to work with GRanges objects
