### 2026-07-09 efa8c6c
Improved argument checks with helper functions and reduced data casting
