+++ man/summarizeVML.Rd
@@ -4,16 +4,16 @@
 \alias{summarizeVML}
 \title{Summarize the methylation states of Variable Methylated Loci (VML)}
 \usage{
-summarizeVML(VML_df, methylation_data)
+summarizeVML(VML, methylation_data)
 }
 \arguments{
-\item{VML_df}{A GRanges-like data frame. Must contain the following columns:
-"seqnames", "start", "end" and "probes" (containing lists as elements, where
-each contains a vector with the probes constituting the VML). This is the
-"VML" object returned by the \emph{findVML()} function.}
+\item{VML}{GRanges object. Must contain a metadata column named "probes",
+where each element contains a vector with the probes constituting the
+VML.}
 
 \item{methylation_data}{A data frame containing M or B values, with samples
-as columns and probes as rows. Row names must be the CpG probe IDs.}
+as columns and probes as rows. Data is expected to have already passed
+through quality control and cleaning steps.}
 }
 \value{
 A data frame with samples as rows, and VML as columns. The value
@@ -50,7 +50,7 @@ VML <- findVML(
 ## Summarize methylation states of the found VML
 summarized_VML <- summarizeVML(
   # Use only 5 for demonstration purposes
-  VML_df = VML$VML[1:5, ],
+  VML = VML$VML[1:5, ],
   methylation_data = test_methylation_data
 )
 
