### 2026-07-09 1cd4e67
Improved input specification.
