+++ man/medCorVMR.Rd
@@ -14,7 +14,7 @@ VML.}
 
 \item{methylation_data}{A data frame containing M or B values, with samples
 as columns and probes as rows. Data is expected to have already passed
-through quality control and cleaning steps.}
+through quality control and cleaning steps. Rows must be the CpG probe IDs.}
 }
 \value{
 A GRanges object like VML with an extra column per region containing
