### 2026-07-10 594f598
Object returned is now a matrix, which homogeneizes input class in the downstream functions
