### 2026-07-10 19c528c
Update tests with matrix output value
