### 2026-07-10 2593104
Add function to check that vectors are the same for argument check, and an empty list vector used throughout the package
