### 2026-07-10 8cf1aea
Update argument check with internal functions, reduce data casting by working with matrices throughout the package, and removed the iter dependency
