+++ R/selectVariables.R
@@ -123,7 +123,7 @@
 #' ## Find cis SNPs around VML
 #' VML_with_cis_snps <- RAMEN::findCisSNPs(
 #'   # Use only 5 for demonstration purposes
-#'   VML_df = VML$VML[1:5, ],
+#'   VML = VML$VML[1:5, ],
 #'   genotype_information = RAMEN::test_genotype_information,
 #'   distance = 1e6
 #' )
@@ -131,12 +131,12 @@
 #' ## Summarize methylation levels in VML
 #' summarized_methyl_VML <- RAMEN::summarizeVML(
 #'   methylation_data = RAMEN::test_methylation_data,
-#'   VML_df = VML_with_cis_snps
+#'   VML = VML_with_cis_snps
 #' )
 #'
 #' ## Select relevant genotype and environmental variables
 #' selected_vars <- RAMEN::selectVariables(
-#'   VML_df = VML_with_cis_snps,
+#'   VML_wSNPs = VML_with_cis_snps,
 #'   genotype_matrix = RAMEN::test_genotype_matrix,
 #'   environmental_matrix = RAMEN::test_environmental_matrix,
 #'   covariates = RAMEN::test_covariates,
@@ -144,139 +144,79 @@
 #'   seed = 1
 #' )
 #'
-selectVariables <- function(VML_df,
+selectVariables <- function(VML_wSNPs,
                             genotype_matrix,
                             environmental_matrix,
                             covariates = NULL,
                             summarized_methyl_VML,
                             seed = NULL) {
   #### Arguments check ####
-  ## Check that genotype_matrix, environmental_matrix, covariate matrix (in case it is provided) and summarized_methyl_VML have the same samples
-  if (!is.data.frame(summarized_methyl_VML)) stop("Please make sure summarized_methyl_VML is a data frame.")
-  if (!all(rownames(summarized_methyl_VML) %in% colnames(genotype_matrix))) stop("Individual IDs in summarized_methyl_VML do not match individual IDs in genotype_matrix")
-  if (!all(rownames(summarized_methyl_VML) %in% rownames(environmental_matrix))) stop("Individual IDs in summarized_methyl_VML do not match individual IDs in environmental_matrix")
-  if (!is.null(covariates)) {
-    if (!all(rownames(summarized_methyl_VML) %in% rownames(covariates))) stop("Individual IDs in summarized_methyl_VML do not match individual IDs in the covariates matrix")
-  }
-  ## Check that VML_df has index and SNP column
-  if (!all(c("VML_index", "SNP") %in% colnames(VML_df))) stop("Please make sure the VML data frame (VML_df) contains the columns 'SNP' and 'VML_index'.")
-  ## Check that the SNP column on VML_df is a list
-  if (!is.list(VML_df$SNP)) stop("Please make sure the 'SNP' column in VML_df is a column containing lists as values")
-  if (!is.character(VML_df$VML_index)) stop("Please make sure the 'VML_index' column in VML_df is a column of characters")
-  ## Check that genotype, environment and covariates are matrices
-  if (!is.matrix(genotype_matrix)) stop("Please make sure the genotype data is provided as a matrix.")
-  if (!is.null(environmental_matrix)) {
-    if (!is.matrix(environmental_matrix)) stop("Please make sure the environmental data is provided as a matrix.")
-  }
-  if (!is.null(covariates)) {
-    if (!is.matrix(covariates)) stop("Please make sure the covariates data is provided as a matrix.")
+  # Check correct input class
+  argument_check(VML_wSNPs, "GRanges")
[…]
