### 2026-07-10 01747cd
Update functions working with GRanges
