### 2026-07-10 836cfda
Update output as matrix
