+++ man/summarizeVML.Rd
@@ -16,7 +16,7 @@ as columns and probes as rows. Data is expected to have already passed
 through quality control and cleaning steps. Rows must be the CpG probe IDs.}
 }
 \value{
-A data frame with samples as rows, and VML as columns. The value
+A matrix with samples as rows, and VML as columns. The value
 inside each cell corresponds to the summarized methylation value of said VML
 in the corresponding individual. The column names correspond to the VML_index.
 }
