### 2026-07-10 f3077e5
Improved argument check with internal functions
