+++ R/lmGE.R
@@ -73,16 +73,16 @@
 #'   publication of the relaimpo R package (Grömping, 2006).
 #'
 #' @param selected_variables A data frame obtained with *RAMEN::selectVariables()*.
-#'  This data frame must contain three columns: 'VML_index' with characters of
-#'  an unique ID of each VML; ´selected_genot' and 'selected_env' with the SNPs
-#'  and environmental variables, respectively, that will be used for fitting the
-#'   genotype (G), environment (E), additive (G + E) or interaction (G x E)
-#'    models. The columns 'selected_env' and 'selected_genot' must contain lists
-#'     as elements; VML with no environmental or genotype selected variables
-#'     must contain an empty list with NULL, NA , character(0) or "" inside.
+#' This data frame must contain three columns: 'VML_index' with characters of
+#' an unique ID of each VML; ´selected_genot' and 'selected_env' with the SNPs
+#' and environmental variables, respectively, that will be used for fitting the
+#' genotype (G), environment (E), additive (G + E) or interaction (G x E)
+#' models. The columns 'selected_env' and 'selected_genot' must contain lists
+#' as elements; VML with no environmental or genotype selected variables
+#' must contain an empty list (i.e., list(NULL), list(NA), list("") or
+#' list(character(0)) ).
 #' @param model_selection Which metric to use to select the best model for each
-#'  VML. Supported options are "AIC" or BIC". More information about which one to
-#'   use can be found in the Details section.
+#' VML. Supported options are "AIC" or BIC".
 #' @inheritParams selectVariables
 #' @return A data frame with the following columns:
 #'  - VML_index: The unique ID of the VML
@@ -135,7 +135,7 @@
 #' ## Find cis SNPs around VML
 #' VML_with_cis_snps <- RAMEN::findCisSNPs(
 #'   # Use only 5 for demonstration purposes
-#'   VML_df = VML$VML [1:5, ],
+#'   VML = VML$VML [1:5, ],
 #'   genotype_information = RAMEN::test_genotype_information,
 #'   distance = 1e6
 #' )
@@ -148,7 +148,7 @@
 #'
 #' ## Select relevant genotype and environmental variables
 #' selected_vars <- RAMEN::selectVariables(
-#'   VML_df = VML_with_cis_snps,
+#'   VML_wSNPs = VML_with_cis_snps,
 #'   genotype_matrix = RAMEN::test_genotype_matrix,
 #'   environmental_matrix = RAMEN::test_environmental_matrix,
 #'   covariates = RAMEN::test_covariates,
@@ -178,89 +178,32 @@ lmGE <- function(selected_variables,
   e_r_squared <- gxe_r_squared <- second_winner <- delta_aic <- NULL
   delta_r_squared <- basal_AIC <- basal_rsquared <- delta_bic <- basal_BIC <- NULL
 
-  # Check arguments
-  # Check that genotype_matrix, environmental_matrix, covariate matrix (in case
-  # it is provided) and summarized_methyl_VML have the same samples
-  if (!is.data.frame(summarized_methyl_VML)) stop("Please make sure summarized_methyl_VML is a data frame.")
-  if (!all(rownames(summarized_methyl_VML) %in% colnames(genotype_matrix))) stop("Individual IDs in summarized_methyl_VML do not match individual IDs in genotype_matrix")
-  if (!all(rownames(summarized_methyl_VML) %in% rownames(environmental_matrix))) stop("Individual IDs in summarized_methyl_VML do not match individual IDs in environmental_matrix")
+  #### Check arguments ####
+  # Input has the right structure
+  argument_check(selected_variables, "data.frame")
[…]
