### 2026-07-10 1dc0316
Reduced data casting, improved conditionals, removed iter dependency
