+++ R/lmGE.R
@@ -203,6 +203,8 @@ lmGE <- function(selected_variables,
   finite_numeric_check(environmental_matrix)
   finite_numeric_check(summarized_methyl_VML)
   if (!is.null(covariates)) finite_numeric_check(covariates)
+
+  #### Select best winning model ####
   # Filter VML that have no selected G and no selected E
   no_vars_VML <- selected_variables |>
     dplyr::filter((selected_env %in% empty_lists &
@@ -210,51 +212,31 @@ lmGE <- function(selected_variables,
   selected_variables <- selected_variables |>
     dplyr::filter(!(selected_env %in% empty_lists &
       selected_genot %in% empty_lists))
-
   # Select the winning model
-  winning_models <- foreach::foreach(
-    VML_i = iterators::iter(selected_variables, by = "row"),
-    .combine = "rbind"
-  ) %dopar% { # For every VML
+  winning_models <- foreach::foreach(i = 1:nrow(selected_variables),
+                                     .combine = "rbind") %dopar% { # For every VML
+    #### Prepare data sets ####
     # Create the data frame with all the information for each VML
-    summ_vml_i <- as.matrix(summarized_methyl_VML[, VML_i$VML_index])
+    VML_i = selected_variables[i,]
+    summ_vml_i <- summarized_methyl_VML[, VML_i$VML_index, drop = FALSE]
     colnames(summ_vml_i) <- "DNAme"
     if (!VML_i$selected_env %in% empty_lists) {
-      if (length(VML_i$selected_env[[1]]) == 1) {
-        env_i <- environmental_matrix[rownames(summarized_methyl_VML),
-                                      unlist(VML_i$selected_env)] |>
-          as.matrix()
-        colnames(env_i) <- unlist(VML_i$selected_env)
-      } else {
-        env_i <- environmental_matrix[rownames(summarized_methyl_VML),
-                                      unlist(VML_i$selected_env)]
-      }
+      env_i <- environmental_matrix[rownames(summarized_methyl_VML),
+                                    unlist(VML_i$selected_env),
+                                    drop = FALSE]
     } else {
       env_i <- NULL
     }
     if (!VML_i$selected_genot %in% empty_lists) {
-      if (length(VML_i$selected_genot[[1]]) == 1) {
-        genot_i <- genotype_matrix[unlist(VML_i$selected_genot),
-                                   rownames(summarized_methyl_VML)] |>
-          as.matrix()
-        colnames(genot_i) <- unlist(VML_i$selected_genot)
+      genot_i <- genotype_matrix[unlist(VML_i$selected_genot),
+                                 rownames(summarized_methyl_VML),
+                                 drop = FALSE] |>
+        t()
       } else {
-        genot_i <- genotype_matrix[unlist(VML_i$selected_genot),
-                                   rownames(summarized_methyl_VML)] |>
-          t()
+        genot_i <- NULL
       }
-    } else {
-      genot_i <- NULL
-    }
     if (!is.null(covariates)) {
-      if (ncol(covariates) == 1) {
-        covariates_i <- covariates[rownames(summarized_methyl_VML), ] |>
-          # Match the covariates dataset with the VML information
-          as.matrix()
-        colnames(covariates_i) <- colnames(covariates)
-      } else {
-        covariates_i <- covariates[rownames(summarized_methyl_VML), ]
-      }
+      covariates_i <- covariates[rownames(summarized_methyl_VML), , drop = FALSE]
     }
     full_data_vml_i <- cbind(summ_vml_i, env_i, genot_i, covariates_i)
     colnames(full_data_vml_i) <- make.names(colnames(full_data_vml_i))
@@ -262,7 +244,7 @@ lmGE <- function(selected_variables,
     basal_model_formula <- colnames(covariates) |>
[…]
