### 2026-07-10 bb2a478
Updated tests with errors thrown by argument checks using internal functions
