### 2026-07-10 6182e75
Removed unnecessary conditionals by adding drop=FALSE in matrix subsetting
