### 2026-07-11 b19abd9
Remove iterators, lifecycle and tibble dependencies
