+++ NAMESPACE
@@ -10,4 +10,3 @@ export(summarizeVML)
 importFrom(doRNG,"%dorng%")
 importFrom(foreach,"%do%")
 importFrom(foreach,"%dopar%")
-importFrom(lifecycle,deprecated)
