### 2026-07-11 95f18cf
Improved argument checks with internal functions, removed unnecessary statements when working with matrices
