+++ R/nullDistGE.R
@@ -34,9 +34,9 @@
 #' distribution and reduces the computational time. For further information
 #' please read the RAMEN paper (https://doi.org/10.1186/s13059-025-03864-4).
 #'
-#' @param permutations description
 #' @inheritParams selectVariables
 #' @inheritParams lmGE
+#' @param permutations Numer of permutation analyses to run.
 #'
 #' @return A data frame with the following columns:
 #'  - VML_index: The unique ID of the VML.
@@ -69,7 +69,7 @@
 #' ## Find cis SNPs around VML
 #' VML_with_cis_snps <- findCisSNPs(
 #'   # Use only 5 for demonstration purposes
-#'   VML_df = VML$VML[1:5, ],
+#'   VML = VML$VML[1:5, ],
 #'   genotype_information = test_genotype_information,
 #'   distance = 1e6
 #' )
@@ -77,13 +77,13 @@
 #' ## Summarize methylation levels in VML
 #' summarized_methyl_VML <- summarizeVML(
 #'   methylation_data = test_methylation_data,
-#'   VML_df = VML_with_cis_snps
+#'   VML = VML_with_cis_snps
 #' )
 #'
 #' ## Simulate null distribution of G and E contributions on DNAme variability
 #' ## We will only run one permutation for demonstration purposes
 #' null_dist <- nullDistGE(
-#'   VML_df = VML_with_cis_snps,
+#'   VML_wSNPs = VML_with_cis_snps,
 #'   genotype_matrix = test_genotype_matrix,
 #'   environmental_matrix = test_environmental_matrix,
 #'   summarized_methyl_VML = summarized_methyl_VML,
@@ -93,87 +93,51 @@
 #'   seed = 1,
 #'   model_selection = "AIC"
 #' )
-nullDistGE <- function(VML_df,
+nullDistGE <- function(VML_wSNPs,
                        genotype_matrix,
                        environmental_matrix,
                        summarized_methyl_VML,
-                       permutations = 10,
+                       permutations = 5,
                        covariates = NULL,
                        seed = NULL,
                        model_selection = "AIC") {
   #### Argument check ####
   ## Check that genotype_matrix, environmental_matrix, and covariates (in case
   ## it is provided) have only numeric values and no NA, NaN, Inf values
-  if (!is.data.frame(summarized_methyl_VML)) stop("Please make sure summarized_methyl_VML is a data frame.")
-  if (
-    sum(vapply(genotype_matrix, is.na, FUN.VALUE = logical(1))) > 0 ||
-      sum(vapply(genotype_matrix, is.nan, FUN.VALUE = logical(1))) > 0 ||
-      sum(!vapply(genotype_matrix, is.numeric, FUN.VALUE = logical(1))) > 0 ||
-      sum(vapply(genotype_matrix, is.infinite, FUN.VALUE = logical(1))) > 0
-  ) {
-    stop(
-      "Please make sure the genotype matrix contains only finite numeric values."
-    )
-  }
-  if (
-    sum(vapply(environmental_matrix, is.na, FUN.VALUE = logical(1))) > 0 ||
-      sum(vapply(environmental_matrix, is.nan, FUN.VALUE = logical(1))) > 0 ||
-      sum(!vapply(environmental_matrix, is.numeric, FUN.VALUE = logical(1))) > 0 ||
-      sum(vapply(environmental_matrix, is.infinite, FUN.VALUE = logical(1))) > 0
-  ) {
-    stop(
-      "Please make sure the environmental matrix contains only finite numeric values."
-    )
-  }
-  if (!is.null(covariates)) {
-    if (
-      sum(vapply(covariates, is.na, FUN.VALUE = logical(1))) > 0 ||
-        sum(vapply(covariates, is.nan, FUN.VALUE = logical(1))) > 0 ||
[…]
