### 2026-07-11 fbbcf4a
Split the code so lines are shorter
