+++ R/test_methylation_data.R
@@ -1,6 +1,8 @@
 #' Methylation data matrix example
 #'
-#' Simulated M values of the 3000 probes selected in *test_array_manifest* for 30 individuals. Values were converted from beta values, which were drawn from a bimodal Beta distribution.
+#' Simulated M values of the 3000 probes selected in *test_array_manifest* for
+#' 30 individuals. Values were converted from beta values, which were drawn from
+#' a bimodal Beta distribution.
 #'
 #' @format ## `test_methylation_data`
 #' A data frame with 3,000 rows and 30 columns:
