+++ R/test_covariates.R
@@ -6,6 +6,7 @@
 #' A data frame with 30 rows and 1 column:
 #' \describe{
 #'   \item{*rownames*}{Individual IDs}
-#'   \item{covar1}{Concomitant variable drawn from a normal distribution with mean=0 and sd=1}
+#'   \item{covar1}{Concomitant variable drawn from a normal distribution with
+#'   mean = 0 and sd = 1}
 #' }
 "test_covariates"
