+++ man/test_covariates.Rd
@@ -10,7 +10,8 @@
 A data frame with 30 rows and 1 column:
 \describe{
 \item{\emph{rownames}}{Individual IDs}
-\item{covar1}{Concomitant variable drawn from a normal distribution with mean=0 and sd=1}
+\item{covar1}{Concomitant variable drawn from a normal distribution with
+mean = 0 and sd = 1}
 }
 }
 }
