+++ man/test_genotype_matrix.Rd
@@ -19,6 +19,9 @@ A data frame with 8,539 rows and 30 columns:
 test_genotype_matrix
 }
 \description{
-Genotype matrix example using a gene-dosage model, which encodes the SNPs ordinally depending on the genotype allele charge, such as 2 (AA), 1 (AB) and 0 (BB). Valus were drawn from a binomial distribution with size 2 and probability 0.5.
+Genotype matrix example using a gene-dosage model, which encodes the SNPs
+ordinally depending on the genotype allele charge, such as 2 (AA), 1 (AB) and
+0 (BB). Valus were drawn from a binomial distribution with size 2 and
+probability 0.5.
 }
 \keyword{datasets}

[…]
