### 2026-07-11 698e83f
Updated tests now that functions work with GRanges
