+++ tests/testthat/test-nullDistGE.R
@@ -2,7 +2,7 @@ test_that("nullDistGE output structure is correct", {
   # Set the parallel backend to use 2 workers
   doParallel::registerDoParallel(2)
   null_dist <- RAMEN::nullDistGE(
-    VML_df = VML_cis_snps_test[1:10, ],
+    VML_wSNPs = VML_cis_snps_test[1:5, ],
     genotype_matrix = RAMEN::test_genotype_matrix,
     environmental_matrix = RAMEN::test_environmental_matrix,
     summarized_methyl_VML = summarized_methyl_VML_test,
@@ -14,18 +14,18 @@ test_that("nullDistGE output structure is correct", {
     suppressMessages()
   expect_true(is.data.frame(null_dist))
   expect_equal(ncol(null_dist), 6)
-  expect_equal(nrow(null_dist), nrow(VML_cis_snps_test[1:10, ]) * 2)
+  expect_equal(nrow(null_dist), length(VML_cis_snps_test[1:5, ]) * 2)
 })
 
 test_that("nullDistGE works with BIC", {
   # Set the parallel backend to use 2 workers
   doParallel::registerDoParallel(2)
   null_dist_bic <- RAMEN::nullDistGE(
-    VML_df = VML_cis_snps_test[1:10, ],
+    VML_wSNPs = VML_cis_snps_test[1:2, ],
     genotype_matrix = RAMEN::test_genotype_matrix,
     environmental_matrix = RAMEN::test_environmental_matrix,
     summarized_methyl_VML = summarized_methyl_VML_test,
-    permutations = 2,
+    permutations = 1,
     covariates = RAMEN::test_covariates,
     seed = 1,
     model_selection = "BIC"
@@ -33,13 +33,13 @@ test_that("nullDistGE works with BIC", {
     suppressMessages()
   expect_true(is.data.frame(null_dist_bic))
   expect_equal(ncol(null_dist_bic), 6)
-  expect_equal(nrow(null_dist_bic), nrow(VML_cis_snps_test[1:10, ]) * 2)
+  expect_equal(nrow(null_dist_bic), length(VML_cis_snps_test[1:2, ]) * 1)
 })
 
 test_that("nullDistGE throws errors when expected", {
   expect_error(
     RAMEN::nullDistGE(
-      VML_df = "a",
+      VML_wSNPs = "a",
       genotype_matrix = RAMEN::test_genotype_matrix,
       environmental_matrix = RAMEN::test_environmental_matrix,
       summarized_methyl_VML = summarized_methyl_VML_test,
@@ -49,7 +49,7 @@ test_that("nullDistGE throws errors when expected", {
       model_selection = "AIC"
     ) |>
       suppressMessages(),
-    "Please make sure the VML data frame (VML_df) contains the columns 'SNP' and 'VML_index'.",
+    "Please make sure the input VML_wSNPs belongs to the GRanges class.",
     fixed = TRUE
   )
   # Test error when genotype_matrix has NA
@@ -58,7 +58,7 @@ test_that("nullDistGE throws errors when expected", {
   test_genot_na[1, 1] <- NA
   expect_error(
     RAMEN::nullDistGE(
-      VML_df = VML_cis_snps_test[1:10, ],
+      VML_wSNPs = VML_cis_snps_test[1:10, ],
       genotype_matrix = test_genot_na,
       environmental_matrix = RAMEN::test_environmental_matrix,
       summarized_methyl_VML = summarized_methyl_VML_test,
@@ -68,16 +68,16 @@ test_that("nullDistGE throws errors when expected", {
       model_selection = "AIC"
     ) |>
       suppressMessages(),
-    "Please make sure the genotype matrix contains only finite numeric values.",
[…]
