### 2026-07-11 40ac8ab
Removed stringr dependency, added methods
