### 2026-07-11 77136b7
Removed stringr functions and replaced them with base R
