+++ R/lmGE.R
@@ -173,7 +173,7 @@ lmGE <- function(selected_variables,
                  covariates = NULL,
                  model_selection = "AIC") {
   #### Binding of variables used within the tidyverse framework ####
-  selected_env <- selected_genot <- VML_i <- SNP <- env <- model_group <- NULL
+  selected_env <- selected_genot <- i <- SNP <- env <- model_group <- NULL
   AIC <- BIC <- tot_r_squared <- VML_index <- variables <- g_r_squared <- NULL
   e_r_squared <- gxe_r_squared <- second_winner <- delta_aic <- NULL
   delta_r_squared <- basal_AIC <- basal_rsquared <- delta_bic <- basal_BIC <- NULL
@@ -253,10 +253,10 @@ lmGE <- function(selected_variables,
       ) %do% { # For each SNP
         ### Fit G models
         model_g <- stats::lm(data = as.data.frame(full_data_vml_i),
-                             formula = stringr::str_glue("DNAme ~ ",
-                                                         make.names(SNP),
-                                                         " + ",
-                                                         basal_model_formula))
+                             formula = paste("DNAme ~",
+                                             make.names(SNP),
+                                             "+",
+                                             basal_model_formula))
 
         # Create data frame structure for the results
         model_g_df <- data.frame(model_group = "G")
@@ -276,12 +276,12 @@ lmGE <- function(selected_variables,
           ) %do% {
             # Fit G + E
             model_ge <- stats::lm(data = as.data.frame(full_data_vml_i),
-                                  formula = stringr::str_glue("DNAme ~ ",
-                                                              make.names(SNP),
-                                                              " + ",
-                                                              make.names(env),
-                                                              " + ",
-                                                              basal_model_formula)
+                                  formula = paste("DNAme ~",
+                                                  make.names(SNP),
+                                                  "+",
+                                                  make.names(env),
+                                                  "+",
+                                                  basal_model_formula)
                                   )
 
             # Create data frame structure for the results
@@ -295,15 +295,16 @@ lmGE <- function(selected_variables,
             model_ge_df$tot_r_squared <- summary(model_ge)$r.squared
             # Fit GxE
             model_gxe <- stats::lm(data = as.data.frame(full_data_vml_i),
-                                   formula = stringr::str_glue("DNAme ~ ",
-                                                               make.names(SNP),
-                                                               " + ",
-                                                               make.names(env),
-                                                               " + ",
-                                                               make.names(SNP),
-                                                               "*", make.names(env),
-                                                               " + ",
-                                                               basal_model_formula)
+                                   formula = paste("DNAme ~ ",
+                                                   make.names(SNP),
+                                                   " + ",
+                                                   make.names(env),
+                                                   " + ",
+                                                   make.names(SNP),
+                                                   "*", make.names(env),
+                                                   " + ",
+                                                   basal_model_formula,
+                                                   sep = "")
                                    )
             # Create data frame structure for the results
             model_gxe_df <- data.frame(model_group = "GxE")
@@ -339,10 +340,10 @@ lmGE <- function(selected_variables,
       ) %do% {
         # Fit E models
         model_e <- stats::lm(data = as.data.frame(full_data_vml_i),
-                             formula = stringr::str_glue("DNAme ~ ",
-                                                         make.names(env),
-                                                         " + ",
-                                                         basal_model_formula)
+                             formula = paste("DNAme ~",
+                                             make.names(env),
[…]
