### 2026-07-11 465d824
Specified packages for missing functions
