### 2026-07-11 cf5ca42
Updated examples
