+++ man/findCisSNPs.Rd
@@ -37,7 +37,7 @@ the ones in the genotype information are from the same genome build.
 }
 \examples{
 ## Find VML in test data
-VML <- RAMEN::findVML(
+VML <- findVML(
   methylation_data = RAMEN::test_methylation_data,
   array_manifest = "IlluminaHumanMethylationEPICv1",
   cor_threshold = 0,
@@ -48,7 +48,7 @@ VML <- RAMEN::findVML(
 )
 ## Find cis SNPs around VML
 # Use only 5 for demonstration purposes
-VML_with_cis_snps <- RAMEN::findCisSNPs(
+VML_with_cis_snps <- findCisSNPs(
   VML = VML$VML[1:5, ],
   genotype_information = RAMEN::test_genotype_information,
   distance = 1e6
