### 2026-07-11 5836a77
Updated global variable
