### 2026-07-13 61d190d
Addressed lintr comments
