+++ README.md
@@ -55,14 +55,17 @@ You can install the latest version of RAMEN from
 
 ``` r
 ## Install dependencies
-# install.packages("devtools")
 # install.packages("BiocManager")
 # BiocManager::install("S4Vectors")
 # BiocManager::install("IRanges")
 # BiocManager::install("GenomicRanges")
+## If using any of these Illumina microarrays, pick one: 
+# BiocManager::install("IlluminaHumanMethylation450kanno.ilmn12.hg19")
+# BiocManager::install("IlluminaHumanMethylationEPICanno.ilm10b4.hg19")
+# BiocManager::install("IlluminaHumanMethylationEPICv2anno.20a1.hg38")
 
 ## Install the RAMEN package from GitHub
-BiocManager::install("ErickNavarroD/RAMEN@ropensci-software-review")
+BiocManager::install("ErickNavarroD/RAMEN")
 ```
 
 ## Usage
@@ -136,21 +139,26 @@ VML <- RAMEN::findVML(
 #> Identifying Variable Methylated Regions...
 #> Applying correlation filter to Variable Methylated Regions...
 
-head(VML$VML) # Take a look at the identified VML data frame
-#>   VML_index type seqnames    start      end width strand       probes n_VMPs
-#> 1      VML1  VMR    chr21 10990119 10990903   785      + cg098720....      2
-#> 2      VML2  VMR    chr21 11109021 11109336   316      + cg007508....      2
-#> 3      VML3  VMR    chr21 31799091 31799248   158      + cg245007....      2
-#> 4      VML4  VMR    chr21 32715908 32716792   885      + cg164170....      2
-#> 5      VML5  VMR    chr21 15955548 15955699   152      - cg147721....      2
-#> 6      VML6  VMR    chr21 26573136 26573196    61      - cg111120....      2
-#>   median_correlation
-#> 1          0.6099180
-#> 2          0.6261681
-#> 3          0.7279154
-#> 4          0.6932442
-#> 5          0.8120654
-#> 6          0.6173683
+head(VML$VML) # Take a look at the identified VML GRanges object
+#> GRanges object with 6 ranges and 5 metadata columns:
+#>       seqnames            ranges strand |    n_VMPs                probes
+#>          <Rle>         <IRanges>  <Rle> | <numeric>                <list>
+#>   [1]    chr21 10990119-10990903      + |         2 cg09872009,cg05437132
+#>   [2]    chr21 11109021-11109336      + |         2 cg00750806,cg12301579
+#>   [3]    chr21 31799091-31799248      + |         2 cg24500711,cg07621949
+#>   [4]    chr21 32715908-32716792      + |         2 cg16417027,cg14151498
+#>   [5]    chr21 15955548-15955699      - |         2 cg14772146,cg07412745
+#>   [6]    chr21 26573136-26573196      - |         2 cg11112002,cg23973918
+#>       median_correlation        type   VML_index
+#>                <numeric> <character> <character>
+#>   [1]           0.609918         VMR        VML1
+#>   [2]           0.626168         VMR        VML2
+#>   [3]           0.727915         VMR        VML3
+#>   [4]           0.693244         VMR        VML4
+#>   [5]           0.812065         VMR        VML5
+#>   [6]           0.617368         VMR        VML6
+#>   -------
+#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
 ```
 
 2.  Summarize the regional methylation state of each VML with
@@ -158,7 +166,7 @@ head(VML$VML) # Take a look at the identified VML data frame
 
 ``` r
 summarized_methyl_VML <- RAMEN::summarizeVML(
[…]
