### 2026-07-14 224d74b
Remove roxygen tags from internal function
