### 2026-07-14 b8a4b9e
remove roxygen tag from internal function
