+++ man/map_revmap_names.Rd
@@ -1,32 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/findVML.R
-\name{map_revmap_names}
-\alias{map_revmap_names}
-\title{Map revmap column to probe names after reducing a GenomicRanges object}
-\usage{
-map_revmap_names(positions, manifest_hvp)
-}
-\arguments{
-\item{positions}{A revmap row in the form of a vector}
-
-\item{manifest_hvp}{the manifest of the highly variable probes used in the
-findVML() function with the probes as row names}
-}
-\value{
-a vector with the names of the probes that conform one reduced region
-}
-\description{
-Given a revmap row (e.g. 1 5 6), we map those positions to their
-corresponding probe names (and end up with something like "cg00000029",
-"cg00000158", "cg00000165".This is a helper function of findVML()).
-}
-\examples{
-\dontrun{
-target <- data.frame(row.names = c("a", "b", "c", "d"),
-                     values = c(1, 1, 1, 1))
-query <- c(2, 1)
-
-map_revmap_names(positions = query, manifest_hvp = target)
-## Expected output: c("b", "a")
-}
-}
