### 2026-07-14 0e570b1
Remove trailing spaces
