+++ tests/testthat/test-summarizeVML.R
@@ -18,8 +18,7 @@ test_that("summarizeVML adds VML_index when not present", {
   expect_true(is.matrix(summarized_no_index))
   # All VMLs are included
   expect_true(all(colnames(summarized_no_index) %in% paste0("VML",
-                                                           seq_len(length(VML_no_IDs)))
-  )
+                                                           seq_len(length(VML_no_IDs)))))
   # All individuals are included
   expect_equal(nrow(summarized_no_index), ncol(RAMEN::test_methylation_data))
 })
