### 2026-07-14 0ea03fa
Add guidance on findCisSNPs window
