### 2026-07-14 9e7417a
Add advantages of using regions compared to single probes
