### 2026-07-14 375429f
Add guidance on window size
