### 2026-07-14 ecc9c9d
Removed symbol that caused an error in latex
