### 2026-07-14 1944ec5
Removed a symbol that caused issues in latex
