### 2026-07-14 ab14d50
Use sequential evaluation so that test coverage is accurrate
