### 2026-07-14 91fe724
Fixed bug that was not throwing an error when the number of ultrastable probes in the data set was low
