+++ NEWS.md
@@ -1,4 +1,4 @@
-# RAMEN (development version)
+# RAMEN 2.1.0
 
 # RAMEN 2.0.1 - June 30, 2026
 
