+++ .github/workflows/R-CMD-check.yaml
@@ -24,7 +24,7 @@ jobs:
           - {os: ubuntu-latest,   r: 'devel', http-user-agent: 'release'}
           - {os: ubuntu-latest,   r: 'release'}
           - {os: ubuntu-latest,   r: 'oldrel-1'}
-          - {os: ubuntu-latest,   r: '4.2.0'}
+          - {os: ubuntu-latest,   r: '4.5.0'}
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
