+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 3.8.3
-pkgdown: 2.2.0
+pkgdown: 2.2.1
 pkgdown_sha: ~
 articles:
   RAMEN: RAMEN.html
-last_built: 2026-06-30T21:12Z
+last_built: 2026-07-14T05:20Z
 urls:
   reference: https://ericknavarrod.github.io/RAMEN/reference
   article: https://ericknavarrod.github.io/RAMEN/articles
