+++ llms.txt
@@ -2,16 +2,14 @@
 
 ## Overview
 
-Regional Association of Methylome variability with the Exposome and
-geNome (RAMEN) is an R package which goal is to estimate the
-contribution of genetic variants and environmental exposures to DNA
-methylation variability at a genome-wide scale. RAMEN takes advantage of
-the fact that DNA methylation levels at nearby CpG sites are often
-correlated, and uses this information to identify Variable Methylated
-Loci (VML) from microarray DNA methylation data. Then, using genomic and
-exposomic data, it can identify which model out of the following
-explains best the DNA methylation variability at each VML: genetic (G),
-environmental (E), additive (G+E) or interactive (GxE).
+**Regional Association of Methylome variability with the Exposome and
+geNome** **(RAMEN)** is an R package which goal is to estimate the
+contribution of genetic variants and environmental exposures to loci
+with high DNA methylation (DNAme) variability at a genome-wide scale
+using population data. Characterizing the factors that contribute to
+DNAme variability is important because DNAme is a key epigenetic
+mechanism that regulates gene expression and plays an important role in
+development, disease, and environmental adaptation.
 
 RAMEN provides a Findable, Accesible, Interoperable and Reusable (FAIR)
 workflow to conduct gene-environment contribution analyses to
@@ -21,9 +19,18 @@ of traditional statistical methods and machine learning approaches,
 RAMEN is designed to be computationally efficient and user-friendly,
 allowing researchers to gain insights into the complex interplay between
 genetics, environment and DNA methylation variability. The package
-includes a detailed tutorial, and individual functions that could be
-useful for other applications beyond the gene-environment contribution
-analysis.
+includes a detailed
+[tutorial](https://ericknavarrod.github.io/RAMEN/articles/RAMEN.html),
+and individual functions that could be useful for other applications
+beyond the gene-environment contribution analysis.
+
+RAMEN takes advantage of the fact that DNA methylation levels at nearby
+CpG sites are often correlated, and uses this information to identify
+Variable Methylated Loci (VML) from microarray DNA methylation data.
+Then, integrating genomic and exposomic data, it can identify which
+model out of the following explains best the DNA methylation variability
+at each VML: genetic (G), environmental (E), additive (G+E) or
+interactive (GxE).
 
 ## Installation
 
@@ -32,13 +39,18 @@ You can install the latest version of RAMEN from
 
 ``` r
 
-# install.packages("devtools")
+## Install dependencies
 # install.packages("BiocManager")
 # BiocManager::install("S4Vectors")
 # BiocManager::install("IRanges")
 # BiocManager::install("GenomicRanges")
+## If using any of these Illumina microarrays, pick one: 
+# BiocManager::install("IlluminaHumanMethylation450kanno.ilmn12.hg19")
+# BiocManager::install("IlluminaHumanMethylationEPICanno.ilm10b4.hg19")
+# BiocManager::install("IlluminaHumanMethylationEPICv2anno.20a1.hg38")
 
-pak::pak("ErickNavarroD/RAMEN")
+## Install the RAMEN package from GitHub
+BiocManager::install("ErickNavarroD/RAMEN")
 ```
 
 ## Usage
@@ -90,6 +102,13 @@ library(dplyr)
 #> 
 #>     intersect, setdiff, setequal, union
[…]
