+++ reference/test_covariates.md
@@ -20,5 +20,5 @@ A data frame with 30 rows and 1 column:
 
 - covar1:
 
-  Concomitant variable drawn from a normal distribution with mean=0 and
-  sd=1
+  Concomitant variable drawn from a normal distribution with mean = 0
+  and sd = 1

[…]
