### 2026-07-14 2fff01c
Iterate in the mcols data frame of the GRanges object instead of the GRanges object to fix bug in WindowsOS
