+++ R/summarizeVML.R
@@ -50,7 +50,6 @@ summarizeVML <- function(VML,
   if (!"VML_index" %in% colnames(S4Vectors::mcols(VML))) {
     S4Vectors::mcols(VML)$VML_index <- paste0("VML", seq_len(length(VML)))
   }
-
   if (!all(unique(unlist(VML$probes)) %in% rownames(methylation_data))) {
     warning(paste("Some probes listed in the VML data frame are not found in",
     "the methylation data. Please check that all probes listed in the 'probes'",
@@ -61,14 +60,12 @@ summarizeVML <- function(VML,
   if (!is.list(VML$probes)) {
     stop("Please make sure the 'probes' column in the VML object is a list")
   }
-
-  VML_index <- i <- NULL # To avoid R CMD check notes
-
+  mcols_vml <- S4Vectors::mcols(VML)
+  VML_1 <- i <- NULL # To avoid R CMD check notes
   #### Summarize VML ####
-  summarized_VML <- foreach::foreach(i = VML$VML_index,
+  summarized_VML <- foreach::foreach(i = mcols_vml$VML_index,
                                      .combine = "cbind") %dopar% {
-    probes <- VML[VML$VML_index == i]$probes |>
-      unlist()
+    probes <- unlist(mcols_vml$probes[mcols_vml$VML_index == i])
     subset_meth <-  methylation_data[probes, , drop = FALSE]
     median <- apply(subset_meth, 2, median, na.rm = TRUE)
     matrix(median, ncol = 1, dimnames = list(NULL, i))
