+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.2.1
 pkgdown_sha: ~
 articles:
   RAMEN: RAMEN.html
-last_built: 2026-07-14T05:28Z
+last_built: 2026-07-14T19:15Z
 urls:
   reference: https://ericknavarrod.github.io/RAMEN/reference
   article: https://ericknavarrod.github.io/RAMEN/articles
