+++ articles/RAMEN.md
@@ -798,6 +798,66 @@ selected_variables <- RAMEN::selectVariables(
   summarized_methyl_VML = summarized_methyl_VML,
   seed = 1
 )
+#> Loading required package: stats4
+#> Loading required package: BiocGenerics
+#> Loading required package: generics
+#> 
+#> Attaching package: 'generics'
+#> The following object is masked from 'package:dplyr':
+#> 
+#>     explain
+#> The following objects are masked from 'package:base':
+#> 
+#>     as.difftime, as.factor, as.ordered, intersect, is.element, setdiff,
+#>     setequal, union
+#> 
+#> Attaching package: 'BiocGenerics'
+#> The following object is masked from 'package:dplyr':
+#> 
+#>     combine
+#> The following objects are masked from 'package:stats':
+#> 
+#>     IQR, mad, sd, var, xtabs
+#> The following objects are masked from 'package:base':
+#> 
+#>     anyDuplicated, aperm, append, as.data.frame, basename, cbind,
+#>     colnames, dirname, do.call, duplicated, eval, evalq, Filter, Find,
+#>     get, grep, grepl, is.unsorted, lapply, Map, mapply, match, mget,
+#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
+#>     rbind, Reduce, rownames, sapply, saveRDS, table, tapply, unique,
+#>     unsplit, which.max, which.min
+#> Loading required package: S4Vectors
+#> 
+#> Attaching package: 'S4Vectors'
+#> The following object is masked from 'package:tidyr':
+#> 
+#>     expand
+#> The following objects are masked from 'package:dplyr':
+#> 
+#>     first, rename
+#> The following object is masked from 'package:utils':
+#> 
+#>     findMatches
+#> The following objects are masked from 'package:base':
+#> 
+#>     expand.grid, I, unname
+#> Loading required package: IRanges
+#> 
+#> Attaching package: 'IRanges'
+#> The following objects are masked from 'package:dplyr':
+#> 
+#>     collapse, desc, slice
+#> Loading required package: Seqinfo
+#> Loading required package: Matrix
+#> 
+#> Attaching package: 'Matrix'
+#> The following object is masked from 'package:S4Vectors':
+#> 
+#>     expand
+#> The following objects are masked from 'package:tidyr':
+#> 
+#>     expand, pack, unpack
+#> Loaded glmnet 5.0
 #> Loading required package: rngtools
 ```
 
@@ -1368,48 +1428,52 @@ sessionInfo()
 #> tzcode source: system (glibc)
 #> 
 #> attached base packages:
-#> [1] parallel  stats     graphics  grDevices utils     datasets  methods  
-#> [8] base     
+#> [1] stats4    parallel  stats     graphics  grDevices utils     datasets 
+#> [8] methods   base     
 #> 
 #> other attached packages:
-#>  [1] doRNG_1.8.6.3     rngtools_1.5.2    doParallel_1.0.17 iterators_1.0.14 
-#>  [5] foreach_1.5.2     tidyr_1.3.2       ggplot2_4.0.3     dplyr_1.2.1      
-#>  [9] RAMEN_2.1.0       knitr_1.51        BiocStyle_2.40.0 
+#>  [1] doRNG_1.8.6.3        rngtools_1.5.2       glmnet_5.0          
+#>  [4] Matrix_1.7-5         GenomicRanges_1.64.0 Seqinfo_1.2.0       
+#>  [7] IRanges_2.46.0       S4Vectors_0.50.1     BiocGenerics_0.58.1 
[…]
