### 2026-07-15 5af50ba
Evaluate examples sequentially to fix windows bug in the examples
