+++ R/nullDistGE.R
@@ -54,8 +54,8 @@
 #' @importFrom foreach %do%
 #' @export
 #' @examples
-#' # Set the parallel backend to use 2 workers
-#' doParallel::registerDoParallel(2)
+#' # Evaluate sequentially
+#' foreach::registerDoSEQ()
 #' ## Find VML in test data
 #' VML <- findVML(
 #'   methylation_data = test_methylation_data,
