+++ R/summarizeVML.R
@@ -22,8 +22,8 @@
 #'
 #' @examples
 #' ## Find VML in test data
-#' # Set the parallel backend to use 2 workers
-#' doParallel::registerDoParallel(2)
+#' # Evaluate sequentially
+#' foreach::registerDoSEQ()
 #' VML <- findVML(
 #'   methylation_data = test_methylation_data,
 #'   array_manifest = "IlluminaHumanMethylationEPICv1",
