+++ man/lmGE.Rd
@@ -166,8 +166,8 @@ further information, we suggest the users to read the documentation and
 publication of the relaimpo R package (Grömping, 2006).
 }
 \examples{
-# Set the parallel backend to use 2 workers
-doParallel::registerDoParallel(2)
+# Evaluate sequentially
+foreach::registerDoSEQ()
 ## Find VML in test data
 VML <- RAMEN::findVML(
   methylation_data = RAMEN::test_methylation_data,
