+++ man/findVML.Rd
@@ -147,8 +147,8 @@ them, you can do so in the methylation_data object before running the
 function.
 }
 \examples{
-# Set the parallel backend to use 2 workers
-doParallel::registerDoParallel(2)
+# Evaluate sequentially
+foreach::registerDoSEQ()
 VML <- RAMEN::findVML(
   methylation_data = RAMEN::test_methylation_data,
   array_manifest = "IlluminaHumanMethylationEPICv1",
