+++ man/summarizeVML.Rd
@@ -35,8 +35,8 @@ function can be run as usual.
 }
 \examples{
 ## Find VML in test data
-# Set the parallel backend to use 2 workers
-doParallel::registerDoParallel(2)
+# Evaluate sequentially
+foreach::registerDoSEQ()
 VML <- findVML(
   methylation_data = test_methylation_data,
   array_manifest = "IlluminaHumanMethylationEPICv1",
