+++ man/medCorVMR.Rd
@@ -34,8 +34,8 @@ function can be run as usual. It is recommended to also set
 options(future.globals.maxSize= +Inf).
 }
 \examples{
-# Set the parallel backend to use 2 workers
-doParallel::registerDoParallel(2)
+# Evaluate sequentially
+foreach::registerDoSEQ()
 # Create a VML object
 VML <- GenomicRanges::GRanges(seqnames = c("chr21", "chr21"),
       ranges = IRanges::IRanges(start = c(10861376, 10862171),
