+++ reference/lmGE.md
@@ -200,8 +200,8 @@ the documentation and publication of the relaimpo R package (Grömping,
 ## Examples
 
 ``` r
-# Set the parallel backend to use 2 workers
-doParallel::registerDoParallel(2)
+# Evaluate sequentially
+foreach::registerDoSEQ()
 ## Find VML in test data
 VML <- RAMEN::findVML(
   methylation_data = RAMEN::test_methylation_data,
