+++ reference/medCorVMR.md
@@ -40,8 +40,8 @@ options(future.globals.maxSize= +Inf).
 ## Examples
 
 ``` r
-# Set the parallel backend to use 2 workers
-doParallel::registerDoParallel(2)
+# Evaluate sequentially
+foreach::registerDoSEQ()
 # Create a VML object
 VML <- GenomicRanges::GRanges(seqnames = c("chr21", "chr21"),
       ranges = IRanges::IRanges(start = c(10861376, 10862171),
