+++ news/index.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## RAMEN 2.1.0
+## RAMEN 2.1.0 - July 14, 2026
 
 This minor version presents a couple of changes that improve the
 integration of RAMEN with GenomicRanges objects and Bioconductor: -
