### 2026-07-15 8327861
Add instructions to build the vignette
