### 2026-07-15 7fbe0ca
Add recommendation of M vs B values.
