+++ vignettes/RAMEN.Rmd
@@ -201,6 +201,13 @@ design and the plataform, but the steps generally follow this structure:
   can be done using the ComBat function from the 
   [sva package](https://bioconductor.org/packages/release/bioc/html/sva.html). 
 
+We recommend using M values over Beta values, since M value distributions 
+provide more statistically valid properties for linear modelling, such as being
+approximately homoscedastic, unbounded, and with a lower compression at the
+methylated and unmethylated values (for more details see 
+[Pan Du, *et al.*, BMC Bioinformatics, 2010](https://doi.org/10.1186/1471-2105-11-587)). 
+
+
 ### Exposome data 
 
 The exposome is a highly heterogeneous data set, as it can encompass variables 
