### 2026-08-16 #743 @llrs
Many thanks for the quick reply, the new sections on the vignettes are helpful.
All the changes are great. I have just one last comment:

> I was trying to avoid adding the Illumina microarray annotations as dependencies, since they are a bit heavy and take some time to download. Especially because users would typically use only one of the annotations. But since the tests check that the functions works across different annotations,

Sorry, I assumed you already knew, but if some packages are only needed for testing you can leave them on Suggests, but then you need to make sure tests that need them are only run when these packages are present. With testthat you can use `skip_if_not_installed()`. But adding them as dependency is also ok (but it might make it harder for your users to install the package).
