### 2026-08-17 #743 @ropensci-review-bot
Approved! Thanks @ErickNavarroD for submitting and @llrs, @vafadaru for your reviews! :grin:
