### 2026-08-17 #743 @maurolepore
Amazing! Thanks everyone for your contributions!

@ErickNavarroD please notice the TODO list above and check each item as you complete it.

The issue is officially closed but you're welcome to continue to write comments and `@mention` me for support.

Congratulations!
